% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseFunctions.R
\name{addCohortTables}
\alias{addCohortTables}
\title{addCohortTables}
\usage{
addCohortTables(cdm, cohortDatabaseSchema, cohortTableName)
}
\arguments{
\item{cdm}{cdm object}

\item{cohortDatabaseSchema}{cohort db schema}

\item{cohortTableName}{cohort table name}
}
\description{
addCohortTables
}
