% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{allCovariateSettings}
\alias{allCovariateSettings}
\title{Check that all elements are covariateSettings objects}
\usage{
allCovariateSettings(x)
}
\arguments{
\item{x}{A single \code{covariateSettings} object or a list of such objects.}
}
\value{
\code{TRUE} if \code{x} is a single covariateSettings object or a list of covariateSettings objects.
}
\description{
Check that all elements are covariateSettings objects
}
\keyword{internal}
