% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadDatabaseFiles.R
\name{buildPostgresDatabaseSchema}
\alias{buildPostgresDatabaseSchema}
\title{Create postgres database schema for cohort diagnostics results storage}
\usage{
buildPostgresDatabaseSchema(connectionDetails, schemaName, overwrite = FALSE)
}
\arguments{
\item{connectionDetails}{DatabaseConnector connection details compatible object. 
Must be dbms="postgres"}

\item{schemaName}{Schema to install to}

\item{overwrite}{Bool optionally delete and create entire schema again if it exists}
}
\description{
Builds full schema for Diagnostics Explorer
}
