% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkArg.R
\name{checkArg}
\alias{checkArg}
\title{Internal function to check arguments used throughout the CohortDiagnostics package}
\usage{
checkArg(arg, add)
}
\arguments{
\item{arg}{The argument to check}

\item{add}{An optional checkmate::makeAssertCollection() to add error messages to}
}
\description{
Internal function to check arguments used throughout the CohortDiagnostics package
}
