% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckIfCohortInstantiated.R
\name{checkIfCohortInstantiated}
\alias{checkIfCohortInstantiated}
\title{Checks if a set of cohortId(s) are instantiated in the cohort table}
\usage{
checkIfCohortInstantiated(
  connectionDetails = NULL,
  connection = NULL,
  cohortDatabaseSchema,
  cohortTable,
  cohortIds
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{cohortDatabaseSchema}{Schema name where your cohort table resides. Note that for SQL Server,
this should include both the database and schema name, for example
'scratch.dbo'.}

\item{cohortTable}{Name of the cohort table.}

\item{cohortIds}{Provide a set of cohort IDs to check if instantiated.}
}
\value{
Returns TRUE if all cohortIds are instantiated.
}
\description{
Given a set of one or more cohortIds and a single cohort table, checks if 
all cohortIds in the set are instantiated.
}
