% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Private.R
\name{checkInputFileEncoding}
\alias{checkInputFileEncoding}
\title{Check character encoding of input file}
\usage{
checkInputFileEncoding(fileName)
}
\arguments{
\item{fileName}{The path to the file to check}
}
\value{
Throws an error if the input file does not have the correct encoding.
}
\description{
For its input files, CohortDiagnostics only accepts UTF-8 or ASCII character encoding. This
function can be used to check whether a file meets these criteria.
}
