% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortDiagnostics.R
\docType{package}
\name{CohortDiagnostics-package}
\alias{CohortDiagnostics}
\alias{CohortDiagnostics-package}
\title{CohortDiagnostics: Diagnostics for OHDSI Cohorts}
\description{
CohortDiagnostics is an R utility package for the development and evaluation of phenotype algorithms for OMOP CDM compliant data sets. This package provides a standard, end to end, set of analytics for understanding patient capture including data generation and result exploration through an R Shiny interface. Analytics computed include cohort characteristics, record counts, index event misclassification, captured observation windows and basic incidence proportions for age, gender and calendar year. Through the identification of errors, CohortDiagnostics enables the comparison of multiple candidate cohort definitions across one or more data sources, facilitating reproducible research.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ohdsi.github.io/CohortDiagnostics}
  \item \url{https://github.com/OHDSI/CohortDiagnostics}
  \item Report bugs at \url{https://github.com/OHDSI/CohortDiagnostics/issues}
}

}
\author{
\strong{Maintainer}: Jamie Gilbert \email{gilbert@ohdsi.org}

Authors:
\itemize{
  \item Gowtham Rao
  \item Martijn Schuemie
  \item Patrick Ryan
  \item James Weaver
}

Other contributors:
\itemize{
  \item Observational Health Data Science and Informatics [copyright holder]
}

}
\keyword{internal}
