% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsDataModel.R
\name{combineResultsDataModelCsvFiles}
\alias{combineResultsDataModelCsvFiles}
\title{Combine CSVs that are part of Cohort Diagnostics results data model}
\usage{
combineResultsDataModelCsvFiles(
  inputLocation,
  outputLocation = file.path(inputLocation, "combined"),
  recursive = TRUE,
  reportLocation = NULL
)
}
\arguments{
\item{inputLocation}{Input location for files}

\item{outputLocation}{Output location for files}

\item{recursive}{Search for files recursively}

\item{reportLocation}{(Optional) Output data quality report to this location.}
}
\value{
This function searches for csv files that conforms to the Cohort 
Diagnostics results data model, parses them, does quality checks,
appends files and makes them ready for upload into a dbms. Note:
the files are expected to be in snake_case format.
}
\description{
Combine CSVs that are part of Cohort Diagnostics results data model
}
\examples{
\dontrun{
combineResultsDataModelCsvFiles()
}

}
