% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortCharacterizationDiagnostics.R
\name{compareCohortCharacteristics}
\alias{compareCohortCharacteristics}
\title{Compare cohort characteristics}
\usage{
compareCohortCharacteristics(characteristics1, characteristics2)
}
\arguments{
\item{characteristics1}{Characteristics of the first cohort, as created using the
\code{\link{getCohortCharacteristics}} function.}

\item{characteristics2}{Characteristics of the second cohort, as created using the
\code{\link{getCohortCharacteristics}} function.}
}
\value{
A data frame comparing the characteristics of the two cohorts.
}
\description{
Compare the characteristics of two cohorts, computing the standardized difference of the mean.
}
