% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsPulls.R
\name{compareCovariateValueResult}
\alias{compareCovariateValueResult}
\title{Get cohort covariate comparison (including temporal)}
\usage{
compareCovariateValueResult(
  connection = NULL,
  connectionDetails = NULL,
  targetCohortIds,
  comparatorCohortIds,
  databaseIds,
  minProportion = 0.01,
  maxProportion = 1,
  isTemporal = TRUE,
  timeIds = NULL,
  resultsDatabaseSchema = NULL
)
}
\arguments{
\item{connection}{(optional) An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{connectionDetails}{(optional) An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{targetCohortIds}{A vector of one or more Cohort Ids.}

\item{comparatorCohortIds}{A vector of one or more Cohort Ids.}

\item{databaseIds}{A vector one or more databaseIds to retrieve the results for. This is a 
character field values from the `databaseId` field of the `database` table 
of the results data model.}

\item{isTemporal}{(optional) Get temporal covariate values?}

\item{timeIds}{(optional) Used only if isTemporal = TRUE. Do you want to limit to certain
'time ids'. By default all time ids are returned.}

\item{resultsDatabaseSchema}{(optional) The databaseSchema where the results data model 
of cohort diagnostics is stored. This is only required when 
\code{connectionDetails} or \code{\link[DatabaseConnector]{connect}} 
is provided.}
}
\value{
The function will return a tibble data frame object.
}
\description{
Get cohort covariate value comparison data for cohorts stored in cohort diagnostics results data
model. The output of this function may be used, together with covariate ref
to create tables/plots regarding a cohort characteristics comparison diagnostics. 
Because of the large volume of covariates, this function allows to filter range of 
covariate_value by providing the minimum and maximum proportion. Its important to note 
that all covariates are expected to output proportions that are between the values 0.0 to 1.0
}
\details{
Note: The output of this function may be used to create plots or tables. This function relies 
on data available in Cohort Diagnostics results data model. The function will use one of the two
methods to connect to the results data model, 1) database mode, and 2) in-memory mode.

Database mode: In this mode, R will look for the results data model in a remote relational dbms. The 
database system should be one of the databases supported by DatabaseConnector package. If 
either \code{connectionDetails} or \code{\link[DatabaseConnector]{connect}} parameters are
populated with an argument during a function call, the connection gets set to database mode. 

In-memory mode: If both \code{connectionDetails} and \code{\link[DatabaseConnector]{connect}} are 
parameters have a NULL argument then query will be in-memory mode. R will now expect and check
for a data frame object available in R's memory, as required for the function. The object should be
an object specificed in Cohort Diagnostics results data model.

Objects in R's memory are expected to follow camelCase naming conventions (see HADES), while objects 
in a relational database system are expected to follow snake-case naming conventions.
}
\examples{
\dontrun{
covariateValue <- compareCovariateValueResult(targetCohortIds = c(342432,432423),
                                              comparatorCohortIds = c(34243, 342432),
                                              databaseIds = c('eunomia', 'hcup'))
}

}
