% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortComparisonDiagnostics.R
\name{computeCohortOverlap}
\alias{computeCohortOverlap}
\title{Given two sets of cohorts get relationships between the cohorts.}
\usage{
computeCohortOverlap(
  connectionDetails = NULL,
  connection = NULL,
  cohortDatabaseSchema,
  cohortTable = "cohort",
  targetCohortIds,
  comparatorCohortIds,
  batchSize = 200
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{cohortDatabaseSchema}{Schema name where your cohort table resides. Note that for SQL Server,
this should include both the database and schema name, for example
'scratch.dbo'.}

\item{cohortTable}{Name of the cohort table.}

\item{targetCohortIds}{A list of cohort ids to be used as target cohorts.}

\item{comparatorCohortIds}{A list of cohort ids to be used as comparator cohorts.}

\item{batchSize}{{Optional, default set to 200} If running diagnostics on larget set
of cohorts, this function allows you to batch them into chunks that run 
as a batch.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not truly support temp tables. To emulate temp 
tables, provide a schema with write privileges where temp tables can be created.}
}
\description{
Given two cohorts, get data to compare relationships between the cohorts - such as overlap, temporal 
relationships between cohort start date(s). 
Note: only the first occurrence of subject_id in the cohort is used.
}
