% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortRelationship.R
\name{computeCohortTemporalRelationship}
\alias{computeCohortTemporalRelationship}
\title{Given a set of cohorts get relationships between the cohorts.}
\usage{
computeCohortTemporalRelationship(
  connectionDetails = NULL,
  connection = NULL,
  cohortDatabaseSchema,
  cohortTable = "cohort",
  cohortIds,
  batchSize = 200
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{cohortDatabaseSchema}{Schema name where your cohort table resides. Note that for SQL Server,
this should include both the database and schema name, for example
'scratch.dbo'.}

\item{cohortTable}{Name of the cohort table.}

\item{cohortIds}{A vector of one or more Cohort Ids.}

\item{batchSize}{{Optional, default set to 200} If running diagnostics on larget set
of cohorts, this function allows you to batch them into chunks that run 
as a batch.}
}
\description{
Given a set of cohorts, get temporal relationships between the 
cohort_start_date of the cohorts.
}
