% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortDefinitionSet.R
\name{createCohortDefinitionSet}
\alias{createCohortDefinitionSet}
\title{Create a cohort definition set.}
\usage{
createCohortDefinitionSet(cohortJsonFolder)
}
\arguments{
\item{cohortJsonFolder}{The folder containing the cohort Json files.}
}
\value{
a data frame containing the cohort definitions
}
\description{
This method will create a data frame containing the cohort definitions. The data frame contains the
cohort id, name, the sql and the json. The sql is generated from the JSON.
}
