% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortConstruction.R
\name{createCohortTable}
\alias{createCohortTable}
\title{Create cohort table(s)}
\usage{
createCohortTable(connectionDetails, cohortDatabaseSchema,
  cohortTable = "cohort", createInclusionStatsTables = FALSE,
  resultsDatabaseSchema = cohortDatabaseSchema,
  cohortInclusionTable = paste0(cohortTable, "_inclusion"),
  cohortInclusionResultTable = paste0(cohortTable, "_inc_result"),
  cohortInclusionStatsTable = paste0(cohortTable, "_inc_stats"),
  cohortSummaryStatsTable = paste0(cohortTable, "_sum_stats"))
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package.}

\item{cohortDatabaseSchema}{Schema name where your cohort table will be created.
Note that for SQL Server, this should include both the database and
schema name, for example 'scratch.dbo'.}

\item{cohortTable}{Name of the cohort table.}
}
\description{
This function first extracts all concept sets used in a cohort definition. Then, for each concept set
the concept found in the CDM database the contributing source codes are identified.
}
