% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConceptCountsTable.R
\name{createConceptCountsTable}
\alias{createConceptCountsTable}
\title{createConceptCountsTable}
\usage{
createConceptCountsTable(
  connectionDetails = NULL,
  connection = NULL,
  cdmDatabaseSchema,
  tempEmulationSchema = NULL,
  conceptCountsDatabaseSchema,
  conceptCountsTable = "concept_counts",
  conceptCountsTableIsTemp = FALSE,
  removeCurrentTable = TRUE
)
}
\arguments{
\item{connectionDetails}{database connection details}

\item{connection}{database connection}

\item{cdmDatabaseSchema}{CDM schema}

\item{tempEmulationSchema}{schema to emulate temp table}

\item{conceptCountsDatabaseSchema}{schema name for the concept counts table}

\item{conceptCountsTable}{table name}

\item{conceptCountsTableIsTemp}{boolean to indicate if it should be a temporary table}

\item{removeCurrentTable}{if the current table should be removed}
}
\description{
Create a table containing the concept counts.
}
