% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCountsTable.R
\name{createCountsTable}
\alias{createCountsTable}
\title{Function to create the counts table}
\usage{
createCountsTable(connection, cdmDatabaseSchema, writeDatabaseSchema)
}
\arguments{
\item{connection}{Connection to tha database created with DatabaseConnector}

\item{cdmDatabaseSchema}{Schema where cdm is stored}

\item{writeDatabaseSchema}{Schema where table should be created}
}
\description{
Function to create the counts table
}
