% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiagnosticsContext.R
\name{createDiagnosticsContext}
\alias{createDiagnosticsContext}
\title{Create a diagnostics context}
\usage{
createDiagnosticsContext(
  connectionDetails = NULL,
  connection = NULL,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTable = "cohort",
  cohortTableNames = CohortGenerator::getCohortTableNames(cohortTable = cohortTable),
  vocabularyDatabaseSchema = cdmDatabaseSchema,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  databaseId,
  databaseName = NULL,
  databaseDescription = NULL,
  exportFolder,
  minCellCount = 5,
  incremental = FALSE,
  incrementalFolder = file.path(exportFolder, "incremental"),
  cdmVersion = 5
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package.}

\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be used instead of \code{connectionDetails}.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{cohortDatabaseSchema}{Schema name where intermediate data can be stored. Will be used to read
cohorts from. For SQL Server, this should include both the database
and schema name, for example 'scratch.dbo'.}

\item{cohortTable}{The name of the table that contains the cohort definitions.}

\item{cohortTableNames}{An optional list of cohort table names used by CohortGenerator.}

\item{vocabularyDatabaseSchema}{Schema name where your OMOP vocabulary resides. This is commonly the
same as cdmDatabaseSchema. Note that for SQL Server, this should
include both the database and schema name, for example
'vocabulary.dbo'.}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not truly support
temp tables. To emulate temp tables, provide a schema with write
privileges where temp tables can be created.}

\item{databaseId}{A short string for identifying the database (e.g. 'Synpuf').}

\item{databaseName}{The full name of the database.}

\item{databaseDescription}{A short description of the database.}

\item{exportFolder}{The folder where the output will be exported to.}

\item{minCellCount}{The minimum cell count for fields containing person counts or fractions.}

\item{incremental}{Create only cohort diagnostics that haven't been created before?}

\item{incrementalFolder}{If \code{incremental = TRUE}, specify a folder where records are kept.}

\item{cdmVersion}{The version of the CDM (default is 5).}
}
\value{
A DiagnosticsContext object (S3 list)
}
\description{
Creates a context object that holds shared resources for running diagnostics.
This object can be passed to individual diagnostic functions.
}
