% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shiny.R
\name{createDiagnosticsExplorerZip}
\alias{createDiagnosticsExplorerZip}
\title{Create publishable shiny zip}
\usage{
createDiagnosticsExplorerZip(
  outputZipfile = file.path(getwd(), "DiagnosticsExplorer.zip"),
  sqliteDbPath = "MergedCohortDiagnosticsData.sqlite",
  shinyDirectory = system.file(file.path("shiny", "DiagnosticsExplorer"), package =
    "CohortDiagnostics"),
  overwrite = FALSE
)
}
\arguments{
\item{outputZipfile}{The output path for the zip file}

\item{sqliteDbPath}{Merged Cohort Diagnostics sqlitedb created with \code{\link{createMergedResultsFile}}}

\item{shinyDirectory}{(optional) Path to the location where the shiny code is stored. By default,
this is the package root}

\item{overwrite}{If the zip file already exists, overwrite it?}
}
\description{
A utility designed for creating a published zip of a shiny app with an sqlite database.
Designed for sharing projects on servers like data.ohdsi.org.

Takes the shiny code from the R project and adds an sqlite file to a zip archive.
Uncompressed cohort diagnostics sqlite databases can become large very quickly.
}
