% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shiny.R
\name{createMergedResultsFile}
\alias{createMergedResultsFile}
\title{Merge Shiny diagnostics files into sqlite database}
\usage{
createMergedResultsFile(
  dataFolder,
  sqliteDbPath = "MergedCohortDiagnosticsData.sqlite",
  overwrite = FALSE,
  tablePrefix = ""
)
}
\arguments{
\item{dataFolder}{folder where the exported zip files for the diagnostics are stored. Use
the \code{\link{executeDiagnostics}} function to generate these zip files.
Zip files containing results from multiple databases may be placed in the same
folder.}

\item{sqliteDbPath}{Output path where sqlite database is placed}

\item{overwrite}{(Optional) overwrite existing sqlite lite db if it exists.}

\item{tablePrefix}{(Optional) string to insert before table names (e.g. "cd_") for database table names}
}
\description{
This function combines diagnostics results from one or more databases into a single file. The result is an
sqlite database that can be used as input for the Diagnostics Explorer Shiny app.

It also checks whether the results conform to the results data model specifications.
}
