% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shiny.R
\name{deployPositConnectApp}
\alias{deployPositConnectApp}
\title{Rsconnect deploy}
\usage{
deployPositConnectApp(
  appName,
  appDir = tempfile(),
  sqliteDbPath = "MergedCohortDiagnosticsData.sqlite",
  shinyDirectory = system.file(file.path("shiny", "DiagnosticsExplorer"), package =
    "CohortDiagnostics"),
  connectionDetails = NULL,
  shinyConfigPath = NULL,
  resultsDatabaseSchema = NULL,
  vocabularyDatabaseSchemas = resultsDatabaseSchema,
  tablePrefix = "",
  cohortTableName = "cohort",
  databaseTableName = "database",
  port = 80,
  useRenvironFile = FALSE,
  ...
)
}
\arguments{
\item{appName}{string name to call app - should be unique on posit connect server}

\item{appDir}{optional - directory to use to copy files for deployment. If you use a consistent dir
other internal options can change.}

\item{sqliteDbPath}{Path to merged sqlite file. See \code{\link{createMergedResultsFile}} to create file.}

\item{shinyDirectory}{(optional) Directyory shiny app code lives. Use this if you wish to modify the explorer}

\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package, specifying how to connect to the server where
the CohortDiagnostics results have been uploaded using the
\code{\link{uploadResults}} function.}

\item{shinyConfigPath}{Path to shiny yml configuration file (use instead of sqliteDbPath or connectionDetails object)}

\item{resultsDatabaseSchema}{The schema on the database server where the CohortDiagnostics results
have been uploaded.}

\item{vocabularyDatabaseSchemas}{(optional) A list of one or more schemas on the database server where the vocabulary tables are located.
The default value is the value of the resultsDatabaseSchema. We can provide a list of vocabulary schema
that might represent different versions of the OMOP vocabulary tables. It allows us to compare the impact
of vocabulary changes on Diagnostics. Not supported with an sqlite database.}

\item{tablePrefix}{(Optional)  string to insert before table names (e.g. "cd_") for database table names}

\item{cohortTableName}{(Optional) if cohort table name differs from the standard - cohort (ignores prefix if set)}

\item{databaseTableName}{(Optional) if database table name differs from the standard - database (ignores prefix if set)}

\item{port}{(optional) Only used if \code{runOverNetwork} = TRUE.}

\item{useRenvironFile}{logical - not recommended, store db credentials in .Renviron file}

\item{...}{other parameters passed to rsconnect::deployApp}
}
\description{
Deploy your application to an posit connect platform or shinyapps.io server
}
