% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsDataModel.R
\name{dropDdl}
\alias{dropDdl}
\title{Create DDL that drops all results table}
\usage{
dropDdl(packageName, packageVersion, modelVersion, specification)
}
\arguments{
\item{packageName}{The name of the R package whose output model we are documenting.}

\item{packageVersion}{The version number of cohort diagnostics}

\item{modelVersion}{The version of the results data model}

\item{specification}{The location of the csv file with the high-level results table specification.}
}
\description{
Create DDL that drops all results table
}
