% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunDiagnosticsCdm.R
\name{executeDiagnosticsCdm}
\alias{executeDiagnosticsCdm}
\title{Run Diagnostics for a cohort set}
\usage{
executeDiagnosticsCdm(
  cdm,
  cohortSet,
  exportFolder,
  runAnalysis = 1:9,
  minCellCount = 5
)
}
\arguments{
\item{cdm}{A cdm_reference object create by CDMConnector::cdm_from_con()}

\item{cohortSet}{A cohort set created by CDMConnector::readCohortSet()}

\item{exportFolder}{The folder where the output will be exported to. If this folder does not exist it will be created.}

\item{runAnalysis}{A numeric vector corresponding to the diagnostic analyses should be executed. See Details}

\item{minCellCount}{The minimum cell count for fields contains person counts or fractions.}
}
\description{
This is the main function exported by the CohortDiagnositcs package. It runs the
diagnostic analysis and saves the results to a folder.
}
\details{
runAnalysis argument

Pass a numeric vector matching the following analyses. Negative numbers will be interpreted as exclusions.
Use a numeric vector to select the analyses you want to run. Negative numbers will exclude analyses.
The selection works the same way as vector subsetting in R.
}
\section{Analyses to choose from:}{
\enumerate{
\item InclusionStatistics - Report the number of subjects that match specific inclusion rules in each cohort definition.
\item IncludedSourceConcepts - Report the source concepts observed in the database that are included in a concept set of a cohort.
\item OrphanConcepts - Report the source concepts observed in the database that are not included in a concept set of a cohort, but maybe should be.
\item TimeSeries - Boxplot and a table showing the distribution of time (in days) before and after the cohort index date (cohort start date), and the time between cohort start and end date.
\item VisitContext - Report the relationship between the cohort start date and visits recorded in the database.
\item BreakdownIndexEvents - Report the concepts belonging to the concept sets in the entry event definition that are observed on the index date.
\item IncidenceRate - A graph showing the incidence rate, optionally stratified by age (in 10-year bins), gender, and calendar year.
\item CohortRelationship - Stacked bar graph showing the overlap between two cohorts, and a table listing several overlap statistics.
\item TemporalCohortCharacterization - A table showing temporal cohort characteristics (covariates). These characteristics are captured at specific time intervals before or after cohort start date.
}
}

\examples{
\dontrun{

con <- DBI::dbConnect(duckdb::duckdb(), dbdir = CDMConnector::eunomia_dir())

cdm <- cdmFromCon(con, cdmSchema = "main", writeSchema = "main", cdmName = "eunomia")

cohortDefinitionSet <- CDMConnector::readCohortSet(system.file("cohorts", package = "CohortDiagnostics"))

# run all analyses
executeDiagnosticsCdm(
 cdm,

# skip CohortRelationship and TemporalCohortCharacterization
executeDiagnosticsCdm(
 cdm,
}

}
