% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseFunctions.R
\name{executeSql}
\alias{executeSql}
\title{executeSql}
\usage{
executeSql(
  connection,
  sql,
  profile = FALSE,
  progressBar = !as.logical(Sys.getenv("TESTTHAT", unset = FALSE)),
  reportOverallTime = TRUE,
  errorReportFile = file.path(getwd(), "errorReportSql.txt"),
  runAsBatch = FALSE
)
}
\arguments{
\item{connection}{db connection}

\item{sql}{The SQL to be executed}

\item{profile}{profile}

\item{progressBar}{progress bar}

\item{reportOverallTime}{if overall time should be reported}

\item{errorReportFile}{error report file}

\item{runAsBatch}{run as batch?}
}
\description{
executeSql
}
