% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseFunctions.R
\name{existsTable}
\alias{existsTable}
\title{existsTable}
\usage{
existsTable(connection, databaseSchema, tableName)
}
\arguments{
\item{connection}{db connection}

\item{databaseSchema}{db schema name}

\item{tableName}{the table name}
}
\value{
if given table exists
}
\description{
existsTable
}
