% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportFeatureExtractionOutput.R
\name{exportFeatureExtractionOutput}
\alias{exportFeatureExtractionOutput}
\title{Export Feature Extraction output to csv}
\usage{
exportFeatureExtractionOutput(
  featureExtractionDbCovariateData,
  databaseId,
  incremental = FALSE,
  covariateValueFileName = "covariate_value.csv",
  covariateValueContFileName = "covariate_value_dist.csv",
  covariateRefFileName = "covariate_ref.csv",
  analysisRefFileName = "analysis_ref.csv",
  timeDistributionFileName = NULL,
  timeRefFileName = NULL,
  cohortCounts,
  cutOff = 1e-04,
  minCellCount = 5
)
}
\arguments{
\item{featureExtractionDbCovariateData}{An Andromeda object returned by \code{CohortDiagonstics::runCohortCharacterizationDiagnostics}}

\item{databaseId}{A short string for identifying the database (e.g. 'Synpuf').}

\item{incremental}{Create only cohort diagnostics that haven't been created before?}

\item{covariateValueFileName}{The full path (including file name) for the csv file with covariate value data.
e.g. "covariate_value.csv" or "temporal_covariate_value.csv"}

\item{covariateValueContFileName}{The full path (including file name) for the csv file with covariate value distribution data.
e.g. "covariate_value_dist.csv" or "temporal_covariate_value_dist.csv"}

\item{covariateRefFileName}{The full path (including file name) for the csv file with covariate reference data.
e.g. "covariate_ref.csv" or "temporal_covariate_ref.csv"}

\item{analysisRefFileName}{The full path (including file name) for the csv file with analysis reference data.
e.g. "analysis_ref.csv" or "temporal_analysis_ref.csv"}

\item{timeDistributionFileName}{The full path (including file name) for the csv file with time distribution data.
e.g. "time_distribution.csv"}

\item{timeRefFileName}{The full path (including file name) for the csv file with time reference data.
e.g. "temporal_time_ref.csv"}

\item{cohortCounts}{Output \code{CohortDiagnostics::getCohortCounts}}

\item{cutOff}{Minimum value of the covariate value, below which data is censored.}

\item{minCellCount}{(Optional). Default value = 5. The minimum cell count for fields contains person counts or fractions.}
}
\description{
Exports the output of \code{FeatureExtraction::getDbCovariateData} into CSV.
}
