% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConceptSets.R
\name{extractConceptSetsJsonFromCohortJson}
\alias{extractConceptSetsJsonFromCohortJson}
\title{Extract concept set json from cohort json}
\usage{
extractConceptSetsJsonFromCohortJson(cohortJson)
}
\arguments{
\item{cohortJson}{Complete JSON specification of cohort definition. The standard form 
is generated by \code{WebApi}}
}
\value{
The function will return a tibble data frame object with one row per 
conceptSet id in cohort definition.
}
\description{
Extracts json that corresponds to the conceptset definition in a cohort json definition
}
\examples{
\dontrun{
conceptSetsJson <- extractConceptSetsJsonFromCohortJson(cohortJson = json)
}

}
