% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConceptSets.R
\name{extractConceptSetsSqlFromCohortSql}
\alias{extractConceptSetsSqlFromCohortSql}
\title{Extract concept set sql from cohort generation SQL}
\usage{
extractConceptSetsSqlFromCohortSql(cohortSql)
}
\arguments{
\item{cohortSql}{Complete SQL specification of cohort definition in OHDSI SQL dialect. May
contain parameters designed to be replaced by \code{SqlRender}.
The standard form SQL is generated using \code{circe-be} by
\code{WebApi} and \code{Atlas}}
}
\value{
The function will return a tibble data frame object with one row per concept id and
                    concept set combination in cohort definition.
}
\description{
Extracts SQL that corresponds to the conceptset (codeset) part from cohort generation SQL used 
to instantiated conceptSets during cohort construction.
}
\examples{
\dontrun{
conceptSetSql <- extractConceptSetsSqlFromCohortSql(cohortSql = sql)
}

}
