% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiagnosticFinalization.R
\name{finalizeDiagnostics}
\alias{finalizeDiagnostics}
\title{Finalize diagnostics}
\usage{
finalizeDiagnostics(context)
}
\arguments{
\item{context}{A DiagnosticsContext object.}
}
\value{
The DiagnosticsContext object (invisible)
}
\description{
Performs cleanup and exports metadata for cohort diagnostics.
}
