% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaDataDiagnostics.R
\name{findCohortIncludedSourceConcepts}
\alias{findCohortIncludedSourceConcepts}
\title{Check source codes used in a cohort definition}
\usage{
findCohortIncludedSourceConcepts(connectionDetails = NULL,
  connection = NULL, cdmDatabaseSchema, oracleTempSchema = NULL,
  baseUrl = NULL, cohortId = NULL, cohortJson = NULL,
  cohortSql = NULL, byMonth = FALSE, useSourceValues = FALSE)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{oracleTempSchema}{Should be used in Oracle to specify a schema where the user has write
priviliges for storing temporary tables.}

\item{baseUrl}{The base URL for the WebApi instance, for example: "http://server.org:80/WebAPI".}

\item{cohortId}{The ID of the cohort in the WebAPI instance.}

\item{cohortJson}{A characteric string containing the JSON of a cohort definition.}

\item{cohortSql}{The OHDSI SQL representation of the same cohort definition.}

\item{byMonth}{Compute counts by month? If FALSE, only overall counts are computed.}

\item{useSourceValues}{Use the source_value fields to find the codesa used in the data? If not,
this analysis will rely entirely on the source_concept_id fields instead. 
Note that, depending on the source data and ETL, it might be possible for the
source_value fields to contain patient-identifiable information by accident.}
}
\value{
A data frame with source codes, with counts per domain how often the code was encountered
in the CDM.
}
\description{
This function first extracts all concept sets used in a cohort definition. Then, for each concept set
the concept found in the CDM database the contributing source codes are identified.

There are two ways to call this function:
\itemize{
  \item \code{findIncludedSourceCodes(connectionDetails, cdmDatabaseSchema, oracleTempSchema, baseUrl, cohortId)}
  \item \code{findIncludedSourceCodes(connectionDetails, cdmDatabaseSchema, oracleTempSchema, cohortJson, cohortSql)}
}
}
