% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaDataDiagnostics.R
\name{findCohortOrphanConcepts}
\alias{findCohortOrphanConcepts}
\title{Title}
\usage{
findCohortOrphanConcepts(connectionDetails = NULL, connection = NULL,
  cdmDatabaseSchema, oracleTempSchema = NULL, baseUrl = NULL,
  cohortId = NULL, cohortJson = NULL,
  workDatabaseSchema = cdmDatabaseSchema,
  conceptCountsTable = "concept_counts")
}
\arguments{
\item{connectionDetails}{}

\item{connection}{}

\item{cdmDatabaseSchema}{}

\item{oracleTempSchema}{}

\item{baseUrl}{}

\item{cohortId}{}

\item{cohortJson}{}

\item{workDatabaseSchema}{}

\item{conceptCountsTable}{}
}
\description{
Title
}
