% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaDataDiagnostics.R
\name{findOrphanConcepts}
\alias{findOrphanConcepts}
\title{Find source codes that do not roll up to their ancestor}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{oracleTempSchema}{Should be used in Oracle to specify a schema where the user has write
priviliges for storing temporary tables.}

\item{baseUrl}{The base URL for the WebApi instance, for example: "http://server.org:80/WebAPI".}

\item{cohortId}{The ID of the cohort in the WebAPI instance.}

\item{cohortJson}{A characteric string containing the JSON of a cohort definition.}

\item{cohortSql}{The OHDSI SQL representation of the same cohort definition.}
}
\value{
A data frame with orhan concepts, with counts per domain how often the code was encountered
in the CDM.
}
\description{
\code{findOrphanConcepts} searches for concepts that should belong to the concept sets in a cohort definition but don't, for
example because of missing source-to-standard concept maps, or erroneous hierachical relationships.

There are two ways to call this function:
\itemize{
  \item \code{findOrphanConcepts(connectionDetails, cdmDatabaseSchema, oracleTempSchema, baseUrl, cohortId)}
  \item \code{findOrphanConcepts(connectionDetails, cdmDatabaseSchema, oracleTempSchema, cohortJson, cohortSql)}
}
}
\details{
Logically, this function performs the following steps for each concept set expression in the cohort definition:

\itemize{
\item  Given the concept set expression, find all included concepts.
\item  Find all names of those concepts, including synonyms, and the names of source concepts that map to them.
\item  Search for concepts (standard and source) that contain any of those names as substring.
\item  Filter those concepts to those that are not in the original set of concepts (i.e. orphans).
\item  Restrict the set of orphan concepts to those that appear in the CDM database and across network concept prevalence (as either source concept or standard concept).
}
}
