% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSourceInformation.R
\name{getCdmDataSourceInformation}
\alias{getCdmDataSourceInformation}
\title{Returns information from CDM source table.}
\usage{
getCdmDataSourceInformation(
  connectionDetails = NULL,
  connection = NULL,
  cdmDatabaseSchema
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}
}
\value{
Returns a data frame from CDM Data source.
}
\description{
Returns CDM source name, description, release date, CDM release date, version
and vocabulary version, where available.
}
