% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shared.R
\name{getCirceRenderedExpression}
\alias{getCirceRenderedExpression}
\title{Returns list with circe generated documentation}
\usage{
getCirceRenderedExpression(
  cohortDefinition,
  cohortName = "Cohort Definition",
  includeConceptSets = FALSE
)
}
\arguments{
\item{cohortDefinition}{An R object (list) with a list representation of the cohort definition expression,
that may be converted to a cohort expression JSON using
RJSONIO::toJSON(x = cohortDefinition, digits = 23, pretty = TRUE)}

\item{cohortName}{Name for the cohort definition}

\item{includeConceptSets}{Do you want to inclued concept set in the documentation}
}
\value{
list object
}
\description{
Returns list with circe generated documentation
}
