% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortCharacterizationDiagnostics.R
\name{getCohortCharacteristics}
\alias{getCohortCharacteristics}
\title{Create characterization of a cohort}
\usage{
getCohortCharacteristics(
  connectionDetails = NULL,
  connection = NULL,
  cdmDatabaseSchema,
  oracleTempSchema = NULL,
  cohortDatabaseSchema = cdmDatabaseSchema,
  cohortTable = "cohort",
  cohortIds,
  cdmVersion = 5,
  covariateSettings,
  batchSize = 100
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{oracleTempSchema}{Should be used in Oracle to specify a schema where the user has write
privileges for storing temporary tables.}

\item{cohortDatabaseSchema}{Schema name where your cohort table resides. Note that for SQL Server,
this should include both the database and schema name, for example
'scratch.dbo'.}

\item{cohortTable}{Name of the cohort table.}

\item{cohortIds}{A vector of cohortIds (1 or more) used to reference the cohort in the cohort
table.}

\item{cdmVersion}{The version of the OMOP CDM. Default 5. (Note: only 5 is supported.)}

\item{covariateSettings}{Either an object of type \code{covariateSettings} as created using one of
the createCovariate functions in the FeatureExtraction package, or a list
of such objects.}

\item{batchSize}{Maximum number of cohorts to characterize at once. A larger batch size will
be quicker, but may run out of resources on the server.}
}
\value{
An Andromeda object with information on the covariates.
}
\description{
Computes features using all drugs, conditions, procedures, etc. observed on or prior to the cohort
index date.
}
