% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortLevelDiagnostics.R, R/ResultsPulls.R
\name{getCohortCounts}
\alias{getCohortCounts}
\title{Count the cohort(s)}
\usage{
getCohortCounts(
  connection = NULL,
  connect = NULL,
  databaseIds,
  resultsDatabaseSchema = NULL
)

getCohortCounts(
  connection = NULL,
  connect = NULL,
  databaseIds,
  resultsDatabaseSchema = NULL
)
}
\arguments{
\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{databaseIds}{A vector one or more databaseIds to retrieve the results for. This is a character 
field value from the databaseId field in the database table of the results data model.}

\item{resultsDatabaseSchema}{(optional) The databaseSchema where the results data model of cohort diagnostics
is stored. This is only required when \code{connectionDetails} or 
\code{\link[DatabaseConnector]{connect}} is provided.}

\item{cohortDatabaseSchema}{Schema name where your cohort table resides. Note that for SQL Server,
this should include both the database and schema name, for example
'scratch.dbo'.}

\item{cohortTable}{Name of the cohort table.}

\item{cohortIds}{The cohort definition ID(s0 used to reference the cohort in the cohort
table. If left empty, all cohorts in the table will be included.}

\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}
}
\value{
A tibble with cohort counts

The function will return a tibble data frame object.
}
\description{
Computes the subject and entry count per cohort

Get cohort counts data from data stored in cohort diagnostics results data
model. The output of this function may be used to create tables
related to cohort counts diagnostics. This function will look for cohort_count table
in results data model.

Note: this function relies on data available in Cohort Diagnostics results data model. There are
two methods to connect to the resutls data model, database mode and in-memory mode.

Database mode: If either \code{connectionDetails} or \code{\link[DatabaseConnector]{connect}} is 
provided in the function call, the query is set to database mode. In the absence of both 
\code{connectionDetails} and \code{\link[DatabaseConnector]{connect}}, the query will be in-memory 
mode. In in-memory mode, R will expect the data in results data model to be available in R's memory.
In database mode, R will perform a database query. Objects in R's memory are expected to
follow camelCase naming conventions, while objects in dbms are expected to follow snake-case naming
conventions. In database mode, vocabulary tables (if needed) are used from vocabularySchema (which
defaults to resultsDatabaseSchema.). In in-memory mode, vocabulary tables are assumed to in R's 
memory.
}
\examples{
\dontrun{
cohortCounts <- getCohortCounts(resultsDatabaseSchema = resultsDatabaseSchema,
                                databaseIds = c('eunomia', 'hcup'))
}

}
