% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsDataModel.R
\name{getCohortDiagnosticsResultsDataModelSpecification}
\alias{getCohortDiagnosticsResultsDataModelSpecification}
\title{Get specification for Cohort Diagnostics results data model}
\usage{
getCohortDiagnosticsResultsDataModelSpecification()
}
\value{
A tibble data frame object with specifications

#' @examples
\dontrun{
resultsDataModelSpecification <- getResultsDataModelSpecification()
}
}
\description{
Get specification for Cohort Diagnostics results data model
}
