% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCohortRelationship.R
\name{getCohortRelationship}
\alias{getCohortRelationship}
\title{getCohortRelationship}
\usage{
getCohortRelationship(
  connection = NULL,
  cohortDatabaseSchema = NULL,
  tempEmulationSchema = NULL,
  cohortTable = "cohort",
  targetCohortIds,
  comparatorCohortIds,
  relationshipDays
)
}
\arguments{
\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package.}

\item{cohortDatabaseSchema}{Schema name where your cohort table resides. Note that for SQL Server,
this should include both the database and schema name, for example
'scratch.dbo'.}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not truly support temp tables. To emulate temp 
tables, provide a schema with write privileges where temp tables can be created.}

\item{cohortTable}{Name of the cohort table.}

\item{targetCohortIds}{A vector of one or more Cohort Ids for use as target cohorts.}

\item{comparatorCohortIds}{A vector of one or more Cohort Ids for use as feature/comparator cohorts.}

\item{relationshipDays}{A dataframe with two columns startDay and endDay representing periods of time to compute relationship.}
}
\description{
Given a set of cohorts, get temporal relationships between the cohort_start_date of the cohorts.
}
