% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortRelationship.R
\name{getCohortRelationshipDiagnostics}
\alias{getCohortRelationshipDiagnostics}
\title{Given a set of cohorts get relationships between the cohorts.}
\usage{
getCohortRelationshipDiagnostics(
  connectionDetails = NULL,
  connection = NULL,
  tempEmulationSchema = NULL,
  cohortDatabaseSchema = NULL,
  cohortTable = "cohort",
  targetCohortIds,
  comparatorCohortIds,
  relationshipDays
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not truly support temp tables. To emulate temp 
tables, provide a schema with write privileges where temp tables can be created.}

\item{cohortDatabaseSchema}{Schema name where your user has write access (has CRUD privileges). This is the location,
of the cohort tables. This is also the location for the optional table such as 
the unique_concept_id table. 
In incremental mode: It is assumed that the tables in this location have not been modified 
outside this application, and the content was written by this application for current 
project only. A good practice would be to ensure that the database table names are unique 
for the project. e.g. 'EPI101_COHORT', 'EPI101_UNIQUE_CONCEPT_ID' (please try to keep 
length < 22). Another assumption: the content of these tables were written by this application
(i.e. they were not altered outside of this application) and the tables are available during the
life cycle of this project. The local files that are created by the application during incremental
mode such as the 'recordKeepingFile'.
Note that for SQL Server, this should include both the database and schema name, for example
'scratch.dbo'.}

\item{cohortTable}{Name of the cohort table.}

\item{targetCohortIds}{A vector of one or more Cohort Ids for use as target cohorts.}

\item{comparatorCohortIds}{A vector of one or more Cohort Ids for use as feature/comparator cohorts.}

\item{relationshipDays}{A dataframe with two columns startDay and endDay representing periods of time to compute relationship}
}
\description{
Given a set of cohorts, get temporal relationships between the
cohort_start_date of the cohorts.
}
