% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortRelationship.R
\name{getCohortRelationshipIntervalOntology}
\alias{getCohortRelationshipIntervalOntology}
\title{Get Cohort Relationship interval ontology}
\usage{
getCohortRelationshipIntervalOntology()
}
\value{
A tibble describing cohort relationship interval ontology
}
\description{
Get Cohort Relationship interval ontology
}
