% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortConstruction.R
\name{getCohortsJsonAndSql}
\alias{getCohortsJsonAndSql}
\title{Get cohorts JSON and parameterized OHDSI SQL}
\usage{
getCohortsJsonAndSql(
  packageName = NULL,
  cohortToCreateFile = "settings/CohortsToCreate.csv",
  baseUrl = NULL,
  cohortSetReference = NULL,
  cohortIds = NULL
)
}
\arguments{
\item{packageName}{The name of the package containing the cohort definitions. Can be left NULL if 
\code{baseUrl} and \code{cohortSetReference} have been specified.}

\item{cohortToCreateFile}{The location of the cohortToCreate file within the package. Is ignored if 
\code{baseUrl} and \code{cohortSetReference} have been specified. The cohortToCreateFile
must be .csv file that is expected to be read into a dataframe object identical to requirements
for \code{cohortSetReference} argument.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI". Can be left NULL if 
\code{packageName} and \code{cohortToCreateFile} have been specified.}

\item{cohortSetReference}{A data frame with four columns, as described in the details. Can be left NULL if 
\code{packageName} and \code{cohortToCreateFile} have been specified.}

\item{cohortIds}{Optionally, provide a subset of cohort IDs to restrict the
diagnostics to.}
}
\value{
The function will return a R list object with cohort information including specifications 
such as JSON and SQL.
}
\description{
This function may be used to collect a cohorts JSON and OHDSI SQL. Based on whether a
baseUrl is available, the function will collect the specifications from either from
WebApi or a Package.
}
\details{
Currently two ways of executing this function are supported, either 
(1) [Package Mode] embedded in a study package, assuming the cohort definitions are stored in that package using the 
    \code{ROhdsiWebApi::insertCohortDefinitionSetInPackage}, or 
(2) [WebApi Mode] By using a WebApi interface to retrieve the cohort definitions.

When using this function in Package Mode: Use the \code{packageName} and \code{cohortToCreateFile} to specify
the name of the study package, and the name of the cohortToCreate file within that package, respectively

When using this function in WebApi Mode: use the \code{baseUrl} and \code{cohortSetReference} to specify how to 
connect to the WebApi, and which cohorts to fetch, respectively.

Note: if the parameters for both Package Mode and WebApi Mode are provided, then Package mode is preferred.

The \code{cohortSetReference} argument must be a data frame with  the following columns:
\describe{
\item{referentConceptId}{A standard omop concept id that serves as the referant phenotype definition for the cohort Id.}
\item{cohortId}{The cohort Id is the id used to identify  a cohort definition. This is required to be unique. It will be used to create file names. It is recommended to be (referrentConceptId * 1000) + a number between 3 to 999}
\item{webApiCohortId}{Cohort Id in the webApi/atlas instance. It is a required field to run Cohort Diagnostics in WebApi mode. It is discarded in package mode.}
\item{cohortName}{The full name of the cohort. This will be shown in the Shiny app.}
\item{logicDescription}{A human understandable brief description of the cohort definition. This logic does not have to a 
fully specified description of the cohort definition, but should provide enough context to help user understand the
meaning of the cohort definition}
}
}
\examples{
\dontrun{
cohorts <- getCohortsJsonAndSql(packageName = 'cohortDiagnostics',
                                baseUrl = "http://server.org:80/WebAPI")
}

}
