% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortStats.R
\name{getCohortStats}
\alias{getCohortStats}
\title{Get Cohort Inclusion Stats Table Data}
\usage{
getCohortStats(
  connection,
  cohortDatabaseSchema,
  databaseId = NULL,
  snakeCaseToCamelCase = TRUE,
  outputTables = c("cohortInclusionTable", "cohortInclusionResultTable",
    "cohortInclusionStatsTable", "cohortInclusionStatsTable", "cohortSummaryStatsTable",
    "cohortCensorStatsTable"),
  cohortTableNames = getCohortTableNames()
)
}
\arguments{
\item{snakeCaseToCamelCase}{Convert column names from snake case to camel case.}

\item{outputTables}{Character vector. One or more of "cohortInclusionTable", "cohortInclusionResultTable",
"cohortInclusionStatsTable", "cohortInclusionStatsTable", "cohortSummaryStatsTable"
or "cohortCensorStatsTable". Output is limited to these tables. Cannot export, for,
example, the cohort table. Defaults to all stats tables.}
}
\description{
This function returns a data frame of the data in the Cohort Inclusion Tables.
Results are organized in to a list with 5 different data frames:
 * cohortInclusionTable
 * cohortInclusionResultTable
 * cohortInclusionStatsTable
 * cohortSummaryStatsTable
 * cohortCensorStatsTable


These can be optionally specified with the `outputTables`.
}
