% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shared.R
\name{getCohortTemporalRelationshipMatrix}
\alias{getCohortTemporalRelationshipMatrix}
\title{Returns matrix of relationship between target and comparator cohortIds}
\usage{
getCohortTemporalRelationshipMatrix(
  dataSource,
  databaseIds = NULL,
  cohortIds = NULL,
  comparatorCohortIds = NULL,
  relationshipDays = c(-3:3)
)
}
\arguments{
\item{dataSource}{A list object that is the output of \code{createDatabaseDataSource} or \code{createFileDataSource} function. 
This object helps direct the function to query data from the database (created by \code{createDatabaseDataSource})
or a local premerged file (created by \code{createFileDataSource}). Premerged files are output of cohortDiagnostics
compiled into RData using \code{preMergeDiagnosticsFiles}. Database DataSources are data inserted into a remote
database (only a postgres database is supported) with tables created with DDL function \code{createResultsDataModel}
and uploaded using \code{uploadResults}}

\item{databaseIds}{A vector one or more databaseIds to retrieve the results for. This is a 
character field values from the `databaseId` field of the `database` table 
of the results data model.}

\item{cohortIds}{A vector of one or more Cohort Ids.}

\item{relationshipDays}{A vector of integer representing days comparator cohort start to target cohort start}

\item{ComparatorCohortIds}{A vector of one or more Cohort Ids.}
}
\value{
Returns a data frame (tibble)
}
\description{
Given a list of target and comparator cohortIds gets temporal relationship.
}
