% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shared.R
\name{getConceptAncestor}
\alias{getConceptAncestor}
\title{Returns data from concept ancestor table for vector of concept ids}
\usage{
getConceptAncestor(
  dataSource = .GlobalEnv,
  vocabularyDatabaseSchema = NULL,
  conceptId = NULL
)
}
\arguments{
\item{dataSource}{A list object that is the output of \code{createDatabaseDataSource} or \code{createFileDataSource} function. 
This object helps direct the function to query data from the database (created by \code{createDatabaseDataSource})
or a local premerged file (created by \code{createFileDataSource}). Premerged files are output of cohortDiagnostics
compiled into RData using \code{preMergeDiagnosticsFiles}. Database DataSources are data inserted into a remote
database (only a postgres database is supported) with tables created with DDL function \code{createResultsDataModel}
and uploaded using \code{uploadResults}}

\item{vocabularyDatabaseSchema}{Schema name where your OMOP vocabulary data resides. This is 
commonly the same as cdmDatabaseSchema. Note that for 
SQL Server, this should include both the database and
schema name, for example 'vocabulary.dbo'.}

\item{conceptIds}{a vector of concept ids}
}
\value{
Returns a data frame (tibble)
}
\description{
Returns data from concept ancestor table for vector of concept ids
}
