% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConceptCountsTable.R
\name{getConceptCountsTableName}
\alias{getConceptCountsTableName}
\title{getConceptCountsTableName}
\usage{
getConceptCountsTableName(connection, cdmDatabaseSchema)
}
\arguments{
\item{connection}{database connection}

\item{cdmDatabaseSchema}{CDM schema}
}
\value{
the concepts count table name
}
\description{
Get a concept counts table name that is unique for the current database version.
We need to make sure the table is only used if the counts are for the current database.
}
