% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsDataModelQueries.R
\name{getConceptDetails}
\alias{getConceptDetails}
\title{Returns data from concept table of Cohort Diagnostics results data model}
\usage{
getConceptDetails(
  dataSource = .GlobalEnv,
  vocabularyDatabaseSchema = NULL,
  conceptIds
)
}
\arguments{
\item{dataSource}{A list object that is the output of \code{createDatabaseDataSource} function.}

\item{cohortIds}{A vector of one or more Concept Ids.}
}
\value{
Returns a data frame (tibble) with results that conform to concept
table in Cohort Diagnostics results data model.
}
\description{
Returns data from concept table of Cohort Diagnostics results data model
}
