% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shared.R
\name{getConceptMetadata}
\alias{getConceptMetadata}
\title{Returns a metadata for concept ids}
\usage{
getConceptMetadata(
  dataSource,
  databaseIds = NULL,
  vocabularyDatabaseSchema = NULL,
  conceptIds = NULL,
  conceptRelationship = TRUE,
  conceptAncestor = TRUE,
  conceptSynonym = TRUE,
  conceptCount = TRUE
)
}
\arguments{
\item{dataSource}{A list object that is the output of \code{createDatabaseDataSource} or \code{createFileDataSource} function. 
This object helps direct the function to query data from the database (created by \code{createDatabaseDataSource})
or a local premerged file (created by \code{createFileDataSource}). Premerged files are output of cohortDiagnostics
compiled into RData using \code{preMergeDiagnosticsFiles}. Database DataSources are data inserted into a remote
database (only a postgres database is supported) with tables created with DDL function \code{createResultsDataModel}
and uploaded using \code{uploadResults}}

\item{databaseIds}{A vector one or more databaseIds to retrieve the results for. This is a 
character field values from the `databaseId` field of the `database` table 
of the results data model.}

\item{vocabularyDatabaseSchema}{Schema name where your OMOP vocabulary data resides. This is 
commonly the same as cdmDatabaseSchema. Note that for 
SQL Server, this should include both the database and
schema name, for example 'vocabulary.dbo'.}

\item{conceptIds}{(optional) A list of concept ids to limit the metadata result}

\item{conceptRelationship}{Do you want conceptRelationship?}

\item{conceptAncestor}{Do you want conceptAncestor?}

\item{conceptSynonym}{Do you want conceptSynonym?}

\item{conceptCount}{Do you want conceptCount?}
}
\value{
Returns a list of data frames (tibbles)
}
\description{
Returns a metadata for a given list of concept ids that includes concept synonyms,
concept relationship, concept ancestor, concept count per database,
concept cooccurrence on index date per database and cohortId.
}
