% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covariates.R
\name{getDbDefaultCovariateData}
\alias{getDbDefaultCovariateData}
\title{Get default covariate information from the database}
\usage{
getDbDefaultCovariateData(
  connection,
  oracleTempSchema = NULL,
  cdmDatabaseSchema,
  cohortTable = "#cohort_person",
  cohortIds = c(-1),
  cdmVersion = "5",
  rowIdField = "subject_id",
  covariateSettings,
  targetDatabaseSchema,
  targetCovariateTable,
  targetCovariateRefTable,
  targetAnalysisRefTable,
  aggregated = FALSE
)
}
\arguments{
\item{covariateSettings}{Either an object of type \code{covariateSettings} as created using one
of the createCovariate functions, or a list of such objects.}

\item{targetDatabaseSchema}{(Optional) The name of the database schema where the resulting covariates
should be stored.}

\item{targetCovariateTable}{(Optional) The name of the table where the resulting covariates will
be stored. If not provided, results will be fetched to R. The table can be
a permanent table in the \code{targetDatabaseSchema} or a temp table. If
it is a temp table, do not specify \code{targetDatabaseSchema}.}

\item{targetCovariateRefTable}{(Optional) The name of the table where the covariate reference will be stored.}

\item{targetAnalysisRefTable}{(Optional) The name of the table where the analysis reference will be stored.}
}
\description{
Constructs a large default set of covariates for one or more cohorts using data in the CDM schema.
Includes covariates for all drugs, drug classes, condition, condition classes, procedures,
observations, etc.
}
\examples{
\dontrun{
connection <- DatabaseConnector::connect(connectionDetails)
Eunomia::createCohorts(connectionDetails)

results <- getDbDefaultCovariateData(
  connection = connection,
  cdmDatabaseSchema = "main",
  cohortTable = "cohort",
  covariateSettings = createDefaultCovariateSettings(),
  targetDatabaseSchema = "main",
  targetCovariateTable = "ut_cov",
  targetCovariateRefTable = "ut_cov_ref",
  targetAnalysisRefTable = "ut_cov_analysis_ref"
)
}

}
