% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseFunctions.R
\name{CDMConnector::dbms}
\alias{CDMConnector::dbms}
\title{CDMConnector::dbms}
\usage{
CDMConnector::dbms(connection)
}
\arguments{
\item{connection}{db connection}
}
\value{
dbms
}
\description{
CDMConnector::dbms
}
