% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsDataModel.R
\name{getDefaultVocabularyTableNames}
\alias{getDefaultVocabularyTableNames}
\title{Get a list of vocabulary table names}
\usage{
getDefaultVocabularyTableNames()
}
\value{
Get a list of vocabulary table names in results data model
}
\description{
Get a list of vocabulary table names
}
