% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shared.R
\name{getFeatureExtractionCharacterization}
\alias{getFeatureExtractionCharacterization}
\title{Returns cohort characterization output of feature extraction}
\usage{
getFeatureExtractionCharacterization(
  dataSource = .GlobalEnv,
  cohortIds = NULL,
  databaseIds = NULL,
  minThreshold = 0.01
)
}
\arguments{
\item{dataSource}{A list object that is the output of \code{createDatabaseDataSource} or \code{createFileDataSource} function. 
This object helps direct the function to query data from the database (created by \code{createDatabaseDataSource})
or a local premerged file (created by \code{createFileDataSource}). Premerged files are output of cohortDiagnostics
compiled into RData using \code{preMergeDiagnosticsFiles}. Database DataSources are data inserted into a remote
database (only a postgres database is supported) with tables created with DDL function \code{createResultsDataModel}
and uploaded using \code{uploadResults}}

\item{cohortIds}{A vector of one or more Cohort Ids.}

\item{databaseIds}{A vector one or more databaseIds to retrieve the results for. This is a 
character field values from the `databaseId` field of the `database` table 
of the results data model.}

\item{minThreshold}{Do you want to set the minimum threshold for db extraction}
}
\value{
Returns a list object with covariateValue, covariateValueDist,
covariateRef, analysisRef output of feature extraction along with
concept information.
}
\description{
Returns a list object with covariateValue, covariateValueDist,
covariateRef, analysisRef output of feature extraction.
}
