% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetIncidenceProportionData.R
\name{GetIncidenceProportionData}
\alias{GetIncidenceProportionData}
\title{GetIncidenceProportionData}
\usage{
GetIncidenceProportionData(connectionDetails, cohortDatabaseSchema,
  cohortTable, cdmDatabaseSchema, firstOccurrenceOnly = TRUE,
  minObservationTime = 365, cohortDefinitionId, workFolder = NULL)
}
\arguments{
\item{connectionDetails}{The connection details to your database server}

\item{cohortDatabaseSchema}{The database name where your phenotype is instantiated as a standard cohort table.}

\item{cohortTable}{The table name where your phenotype is instantiated as a standard cohort table.}

\item{cdmDatabaseSchema}{The name of your CDM source}

\item{cohortDefinitionId}{Cohort ID}

\item{workFolder}{Optional: directory where time series data to be saved as ipData.rds}
}
\value{
A list of 7 dataframe objects with option to save as an *.rds file.
}
\description{
Returns yearly incidence proportion time series data stratified by age and gender
}
\details{
Returns a list of 7 dataframes of cohort count, database count, and 
incidence proportion per 1000 persons of cohort entry with the following
stratifications: 1) no stratification, 2) gender stratification, 3) age (10-year)
stratification, 4) calendar year and age (10-year) stratification, 5) calendar year and
gender stratification, 6) calendar year, age (10-year), and gender stratification
with option to save dataframes.
}
