% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IncidenceRates.R
\name{getIncidenceRate}
\alias{getIncidenceRate}
\title{Compute incidence rate for a cohort}
\usage{
getIncidenceRate(
  connectionDetails = NULL,
  connection = NULL,
  cohortDatabaseSchema,
  cohortTable,
  cdmDatabaseSchema,
  cdmVersion = 5,
  oracleTempSchema = oracleTempSchema,
  firstOccurrenceOnly = TRUE,
  washoutPeriod = 365,
  cohortId
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{cohortDatabaseSchema}{Schema name where your cohort table resides. Note that for SQL Server,
this should include both the database and schema name, for example
'scratch.dbo'.}

\item{cohortTable}{Name of the cohort table.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{cdmVersion}{The version of the OMOP CDM. Default 5. (Note: only 5 is supported.)}

\item{oracleTempSchema}{Should be used in Oracle to specify a schema where the user has write
privileges for storing temporary tables.}

\item{firstOccurrenceOnly}{Use only the first occurrence of the cohort per person?}

\item{washoutPeriod}{The minimum amount of observation time required before the occurrence
of a cohort entry. This is also used to eliminate immortal time from
the denominator.}

\item{cohortId}{The cohort Id used to reference the cohort in the cohort
table.}
}
\value{
Returns a data frame of cohort count, person year count, and incidence rate per 1000 persons
years with the following stratifications: 1) no stratification, 2) gender stratification,
3) age (10-year) stratification, 4) calendar year and age (10-year) stratification, 5) calendar
year and gender stratification, 6) calendar year, age (10-year), and gender stratification with
option to save dataframes.
}
\description{
Returns yearly incidence rate stratified by age and gender
}
