% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortConstruction.R
\name{getInclusionStatistics}
\alias{getInclusionStatistics}
\title{Title}
\usage{
getInclusionStatistics(connectionDetails, resultsDatabaseSchema,
  instantiatedCohortId, cohortTable = "cohort",
  cohortInclusionTable = paste0(cohortTable, "_inclusion"),
  cohortInclusionResultTable = paste0(cohortTable, "_inclusion_result"),
  cohortInclusionStatsTable = paste0(cohortTable, "_inclusion_stats"),
  cohortSummaryStatsTable = paste0(cohortTable, "_inclusion_stats"))
}
\arguments{
\item{connectionDetails}{}

\item{resultsDatabaseSchema}{}

\item{instantiatedCohortId}{}

\item{cohortTable}{}

\item{cohortInclusionTable}{}

\item{cohortInclusionResultTable}{}

\item{cohortInclusionStatsTable}{}

\item{cohortSummaryStatsTable}{}
}
\value{
A list of data frames.
}
\description{
Title
}
