% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortConstruction.R
\name{getInclusionStatisticsFromFiles}
\alias{getInclusionStatisticsFromFiles}
\title{#' Instantiate a cohort
#'
#' @description
#' This function instantiates the cohort in the cohort table. Optionally, the inclusion rule
#' statistics are computed and stored in the inclusion rule statistics tables.
#'
#' @template Connection
#'
#' @template CohortTable
#'
#' @template CohortDef
#'
#' @template TempEmulationSchema
#' 
#' @template OracleTempSchema
#'
#' @template CdmDatabaseSchema
#' 
#' @template VocabularyDatabaseSchema
#'
#' @param cohortId                     The cohort ID used to reference the cohort in the cohort table.
#' @param generateInclusionStats       Compute and store inclusion rule statistics?
#' @param resultsDatabaseSchema        Schema name where the statistics tables reside. Note that for
#'                                     SQL Server, this should include both the database and schema
#'                                     name, for example 'scratch.dbo'.
#' @param cohortInclusionTable         Name of the inclusion table, one of the tables for storing
#'                                     inclusion rule statistics.
#' @param cohortInclusionResultTable   Name of the inclusion result table, one of the tables for
#'                                     storing inclusion rule statistics.
#' @param cohortInclusionStatsTable    Name of the inclusion stats table, one of the tables for storing
#'                                     inclusion rule statistics.
#' @param cohortSummaryStatsTable      Name of the summary stats table, one of the tables for storing
#'                                     inclusion rule statistics.
#'
#' @export
instantiateCohort <- function(connectionDetails = NULL,
                              connection = NULL,
                              cdmDatabaseSchema,
                              oracleTempSchema = NULL,
                              tempEmulationSchema = NULL,
                              cohortDatabaseSchema = cdmDatabaseSchema,
                              cohortTable = "cohort",
                              baseUrl = NULL,
                              cohortJson = NULL,
                              cohortSql = NULL,
                              cohortId = NULL,
                              generateInclusionStats = FALSE,
                              resultsDatabaseSchema = cohortDatabaseSchema,
                              vocabularyDatabaseSchema = cdmDatabaseSchema,
                              cohortInclusionTable = paste0(cohortTable, "_inclusion"),
                              cohortInclusionResultTable = paste0(cohortTable, "_inclusion_result"),
                              cohortInclusionStatsTable = paste0(cohortTable, "_inclusion_stats"),}
\usage{
getInclusionStatisticsFromFiles(
  cohortIds = NULL,
  folder,
  cohortInclusionFile = file.path(folder, "cohortInclusion.csv"),
  cohortInclusionResultFile = file.path(folder, "cohortIncResult.csv"),
  cohortInclusionStatsFile = file.path(folder, "cohortIncStats.csv"),
  cohortSummaryStatsFile = file.path(folder, "cohortSummaryStats.csv"),
  simplify = TRUE
)
}
\description{
#' Instantiate a cohort
#'
#' @description
#' This function instantiates the cohort in the cohort table. Optionally, the inclusion rule
#' statistics are computed and stored in the inclusion rule statistics tables.
#'
#' @template Connection
#'
#' @template CohortTable
#'
#' @template CohortDef
#'
#' @template TempEmulationSchema
#' 
#' @template OracleTempSchema
#'
#' @template CdmDatabaseSchema
#' 
#' @template VocabularyDatabaseSchema
#'
#' @param cohortId                     The cohort ID used to reference the cohort in the cohort table.
#' @param generateInclusionStats       Compute and store inclusion rule statistics?
#' @param resultsDatabaseSchema        Schema name where the statistics tables reside. Note that for
#'                                     SQL Server, this should include both the database and schema
#'                                     name, for example 'scratch.dbo'.
#' @param cohortInclusionTable         Name of the inclusion table, one of the tables for storing
#'                                     inclusion rule statistics.
#' @param cohortInclusionResultTable   Name of the inclusion result table, one of the tables for
#'                                     storing inclusion rule statistics.
#' @param cohortInclusionStatsTable    Name of the inclusion stats table, one of the tables for storing
#'                                     inclusion rule statistics.
#' @param cohortSummaryStatsTable      Name of the summary stats table, one of the tables for storing
#'                                     inclusion rule statistics.
#'
#' @export
instantiateCohort <- function(connectionDetails = NULL,
                              connection = NULL,
                              cdmDatabaseSchema,
                              oracleTempSchema = NULL,
                              tempEmulationSchema = NULL,
                              cohortDatabaseSchema = cdmDatabaseSchema,
                              cohortTable = "cohort",
                              baseUrl = NULL,
                              cohortJson = NULL,
                              cohortSql = NULL,
                              cohortId = NULL,
                              generateInclusionStats = FALSE,
                              resultsDatabaseSchema = cohortDatabaseSchema,
                              vocabularyDatabaseSchema = cdmDatabaseSchema,
                              cohortInclusionTable = paste0(cohortTable, "_inclusion"),
                              cohortInclusionResultTable = paste0(cohortTable, "_inclusion_result"),
                              cohortInclusionStatsTable = paste0(cohortTable, "_inclusion_stats"),
}
