% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseMigrations.R
\name{getMigrationStatus}
\alias{getMigrationStatus}
\title{getMigrationStatus}
\usage{
getMigrationStatus(
  connection = NULL,
  connectionDetails = NULL,
  schema = NULL,
  tablePrefix = ""
)
}
\arguments{
\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{schema}{The schema on the postgres server where the tables will be created.}

\item{tablePrefix}{(Optional)  string to insert before table names (e.g. "cd_") for database table names}
}
\description{
Get set of migrations that are yet to be run
}
