% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shared.R
\name{getOhdsiSqlFromCohortDefinitionExpression}
\alias{getOhdsiSqlFromCohortDefinitionExpression}
\title{Generates and return cohort SQL}
\usage{
getOhdsiSqlFromCohortDefinitionExpression(
  cohortDefinitionExpression,
  generateStats = TRUE
)
}
\arguments{
\item{generateStats}{Do you want generate intermediary inclusion rules tables?}

\item{cohortExpression}{A cohort definition expression as a list object. This is commonly generated by
by reading in a cohort definition json and converting it into a R-object
using for example RJSONIO::fromJSON(digits = 23). Please use minimum digits = 23,
to ensure that big numbers don't loose precision.'}
}
\value{
Returns text (OHDSI SQL)
}
\description{
Given a cohort definition expression, the function generates and return cohort SQL 
in OHDSI SQL form. This maybe be rendered and translated using 
OHDSI R packages like \code{SqlRender::render}
}
