% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConceptIds.R
\name{getOmopVocabularyTables}
\alias{getOmopVocabularyTables}
\title{Get a copy of omop vocabulary as csv}
\usage{
getOmopVocabularyTables(
  connectionDetails = NULL,
  connection = NULL,
  cdmDatabaseSchema,
  conceptIdTable = NULL,
  vocabularyTableNames = c("concept", "conceptAncestor", "conceptClass",
    "conceptRelationship", "conceptSynonym", "domain", "relationship", "vocabulary"),
  exportFolder
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{cdmDatabaseSchema}{databaseSchema where the omop vocabulary files are located.
These are most commonly the same as cdmDatabaseSchema.}

\item{conceptIdTable}{(optional) A table with one column called concept_id (integer) that
contains all the concept_ids to limit the data pull. In the absence 
of this table, the entire vocabulary is pulled down (slow)
vocabulary files. If NULL, all conceptIds are extracted. Please
provide the full name of the vocabulary table e.g.
databaseSchema.conceptIdTable. If it is a temporary table
please use '#conceptIdTable'. Remember, for temporary table
the connection object has to be active.}

\item{vocabularyTableNames}{(optional) A vector of omop vocabulary table names to download.}

\item{exportFolder}{The folder where the output is exported by Cohort Diagnostics.
If this folder does not exist, or does not have the searched file
the function will return an error.}
}
\value{
NULL. The function writes the vocabulary tables into the export folder as csv.
}
\description{
For a given list of conceptId's get a subset of omop vocabulary
of these conceptIds. These are written as csv in the export folder
}
