% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConceptSetDiagnostics.R
\name{getOptimizationRecommendationForConceptSetExpression}
\alias{getOptimizationRecommendationForConceptSetExpression}
\title{given a concept set table, get optimization recommendation}
\usage{
getOptimizationRecommendationForConceptSetExpression(
  conceptSetExpression,
  vocabularyDatabaseSchema = "vocabulary",
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  connectionDetails = NULL,
  connection = NULL
)
}
\arguments{
\item{conceptSetExpression}{An R Object (list) with concept set expression. This maybe generated
by first getting the JSON representation of concept set expression and
converting it to a list using RJSONIO::fromJson(digits = 23)}

\item{vocabularyDatabaseSchema}{Schema name where your OMOP vocabulary data resides. This is 
commonly the same as cdmDatabaseSchema. Note that for 
SQL Server, this should include both the database and
schema name, for example 'vocabulary.dbo'.}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not truly support temp tables. To emulate temp 
tables, provide a schema with write privileges where temp tables can be created.}

\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}
}
\description{
given a concept set table, get optimization recommendation
}
