% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsPulls.R
\name{getPreComputedCovariateValues}
\alias{getPreComputedCovariateValues}
\title{Get cohort covariate (including temporal)}
\usage{
getPreComputedCovariateValues(
  connection = NULL,
  connectionDetails = NULL,
  cohortIds,
  databaseIds,
  minProportion = 0.01,
  maxProportion = 1,
  isTemporal = TRUE,
  timeIds = NULL,
  resultsDatabaseSchema = NULL
)
}
\arguments{
\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{cohortIds}{Cohort Ids to retrieve the characterization data.}

\item{databaseIds}{A vector one or more databaseIds to retrieve the results for. This is a character 
field value from the databaseId field in the database table of the results data model.}

\item{isTemporal}{(optional) Get temporal covariate values?}

\item{timeIds}{(optional) Used only if isTemporal = TRUE. Do you want to limit to certain
'time ids'. By default all time ids are returned.}

\item{resultsDatabaseSchema}{(optional) The databaseSchema where the results data model of cohort diagnostics
is stored. This is only required when \code{connectionDetails} or 
\code{\link[DatabaseConnector]{connect}} is provided.}
}
\value{
The function will return a tibble data frame object.
}
\description{
Get cohort covariate value data from data stored in cohort diagnostics results data
model. The output of this function may be used, together with covariate ref
to create tables/plots regarding a cohorts characteristics diagnostics. 
Because of the large volume of covariates, this function allows to filter range of 
covariate_value by providing the minimum and maximum proportion. Its important to note 
that all covariates are expected to output proportions that are between the values 0.0 to 1.0

Note: this function relies on data available in Cohort Diagnostics results data model. There are
two methods to connect to the results data model, database mode and in-memory mode.

Database mode: If either \code{connectionDetails} or \code{\link[DatabaseConnector]{connect}} is 
provided in the function call, the query is set to database mode. In the absence of both 
\code{connectionDetails} and \code{\link[DatabaseConnector]{connect}}, the query will be in-memory 
mode. In in-memory mode, R will expect the data in results data model to be available in R's memory.
In database mode, R will perform a database query. Objects in R's memory are expected to
follow camelCase naming conventions, while objects in dbms are expected to follow snake-case naming
conventions. In database mode, vocabulary tables (if needed) are used from vocabularySchema (which
defaults to resultsDatabaseSchema.). In in-memory mode, vocabulary tables are assumed to in R's 
memory.
}
\examples{
\dontrun{
covariateValue <- getPreComputedCovariateValues(cohortId = c(342432,432423),
                                          databaseIds = c('eunomia', 'hcup'))
}

}
