% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortConstruction.R
\name{getRecordCountOfInstantiatedCohorts}
\alias{getRecordCountOfInstantiatedCohorts}
\title{Get record counts for a set of cohort}
\usage{
getRecordCountOfInstantiatedCohorts(
  connection = NULL,
  connectionDetails = NULL,
  cohortDatabaseSchema,
  cohortTable,
  cohortIds
)
}
\arguments{
\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{cohortDatabaseSchema}{Schema name where your cohort table resides. Note that for SQL Server,
this should include both the database and schema name, for example
'scratch.dbo'.}

\item{cohortTable}{Name of the cohort table.}

\item{cohortIds}{A vecotr of cohortIds to get counts for.}
}
\value{
A tibble data frame object
}
\description{
This function get record count for a set of cohorts in the cohort table.
}
