% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shared.R
\name{getResultsCohortCoOccurrenceMatrix}
\alias{getResultsCohortCoOccurrenceMatrix}
\title{Returns data for use in cohort co-occurrence matrix}
\usage{
getResultsCohortCoOccurrenceMatrix(
  dataSource,
  targetCohortIds = NULL,
  comparatorCohortIds = NULL,
  databaseIds = NULL,
  startDay = NULL,
  endDay = NULL,
  showPercent = TRUE
)
}
\arguments{
\item{dataSource}{A list object that is the output of \code{createDatabaseDataSource} or \code{createFileDataSource} function. 
This object helps direct the function to query data from the database (created by \code{createDatabaseDataSource})
or a local premerged file (created by \code{createFileDataSource}). Premerged files are output of cohortDiagnostics
compiled into RData using \code{preMergeDiagnosticsFiles}. Database DataSources are data inserted into a remote
database (only a postgres database is supported) with tables created with DDL function \code{createResultsDataModel}
and uploaded using \code{uploadResults}}

\item{databaseIds}{A vector one or more databaseIds to retrieve the results for. This is a 
character field values from the `databaseId` field of the `database` table 
of the results data model.}

\item{startDay}{A vector of days in relation to cohort_start_date of target}

\item{endDay}{A vector of days in relation to cohort_end_date of target}

\item{showPercent}{Return percent instead of raw numbers}

\item{cohortIds}{A vector of one or more Cohort Ids.}

\item{ComparatorCohortIds}{A vector of one or more Cohort Ids.}
}
\value{
Returns a data frame (tibble). Note - the computation is in relation
to first start of target cohort only.
}
\description{
Returns a a data frame (tibble) that shows the percent (optionally number) of subjects
in target cohort that are also in comparator cohort at certain days relative to 
first start date of a subject in target cohort.
}
