% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsDataModelQueries.R
\name{getResultsCohortCountFromCsv}
\alias{getResultsCohortCountFromCsv}
\title{Get cohort counts from csv file with data in results data model format}
\usage{
getResultsCohortCountFromCsv(csvFile, cohortIds, databaseIds)
}
\arguments{
\item{csvFile}{The full path to read the csv file.}

\item{cohortIds}{One or more cohort ids to returns cohort counts}

\item{databaseIds}{One or more database ids to returns cohort counts}
}
\value{
Returns a tibble dataframe
}
\description{
Given a set of one or more cohortIds and one or more databaseIds,
return data from the cohort_count table in results data model
with data filtered to the cohortIds and databaseIds.
}
