% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shared.R
\name{getResultsCohortOverlap}
\alias{getResultsCohortOverlap}
\title{Returns data for use in cohort_overlap}
\usage{
getResultsCohortOverlap(
  dataSource,
  targetCohortIds = NULL,
  comparatorCohortIds = NULL,
  databaseIds = NULL
)
}
\arguments{
\item{dataSource}{A list object that is the output of \code{createDatabaseDataSource} or \code{createFileDataSource} function. 
This object helps direct the function to query data from the database (created by \code{createDatabaseDataSource})
or a local premerged file (created by \code{createFileDataSource}). Premerged files are output of cohortDiagnostics
compiled into RData using \code{preMergeDiagnosticsFiles}. Database DataSources are data inserted into a remote
database (only a postgres database is supported) with tables created with DDL function \code{createResultsDataModel}
and uploaded using \code{uploadResults}}

\item{targetCohortIds}{A vector of cohort ids representing target cohorts}

\item{comparatorCohortIds}{A vector of cohort ids representing comparator cohorts}

\item{databaseIds}{A vector one or more databaseIds to retrieve the results for. This is a 
character field values from the `databaseId` field of the `database` table 
of the results data model.}
}
\value{
Returns data for use in cohort_overlap
}
\description{
Returns data for use in cohort_overlap
}
