% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shared.R
\name{getResultsCompiledOutput}
\alias{getResultsCompiledOutput}
\title{Extract results from cohort diagnostics}
\usage{
getResultsCompiledOutput(dataSource, cohortIds)
}
\arguments{
\item{dataSource}{A list object that is the output of \code{createDatabaseDataSource} or \code{createFileDataSource} function. 
This object helps direct the function to query data from the database (created by \code{createDatabaseDataSource})
or a local premerged file (created by \code{createFileDataSource}). Premerged files are output of cohortDiagnostics
compiled into RData using \code{preMergeDiagnosticsFiles}. Database DataSources are data inserted into a remote
database (only a postgres database is supported) with tables created with DDL function \code{createResultsDataModel}
and uploaded using \code{uploadResults}}

\item{cohortIds}{A vector of one or more Cohort Ids.}
}
\value{
list of R objects that may then be converted to reports
}
\description{
Extract results from cohort diagnostics results data model for vector of cohort ids
}
