% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shared.R
\name{getResultsConceptCooccurrence}
\alias{getResultsConceptCooccurrence}
\title{Returns concept cooccurrence for a list of cohortIds and databaseIds combinations}
\usage{
getResultsConceptCooccurrence(
  dataSource,
  databaseIds = NULL,
  cohortIds = NULL,
  conceptIds = NULL
)
}
\arguments{
\item{dataSource}{A list object that is the output of \code{createDatabaseDataSource} or \code{createFileDataSource} function. 
This object helps direct the function to query data from the database (created by \code{createDatabaseDataSource})
or a local premerged file (created by \code{createFileDataSource}). Premerged files are output of cohortDiagnostics
compiled into RData using \code{preMergeDiagnosticsFiles}. Database DataSources are data inserted into a remote
database (only a postgres database is supported) with tables created with DDL function \code{createResultsDataModel}
and uploaded using \code{uploadResults}}

\item{databaseIds}{A vector one or more databaseIds to retrieve the results for. This is a 
character field values from the `databaseId` field of the `database` table 
of the results data model.}

\item{cohortIds}{A vector of one or more Cohort Ids.}

\item{conceptId}{A vector of one or more Concept Ids.}
}
\value{
Returns a data frame (tibble)
}
\description{
Given a list of cohortIds, databaseIds combinations the function returns
precomputed concept cooccurrence conceptIds for the combination.
}
