% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shared.R
\name{getResultsConceptCount}
\alias{getResultsConceptCount}
\title{Returns data from concept_count table of Cohort Diagnostics results data model}
\usage{
getResultsConceptCount(
  dataSource,
  databaseIds = NULL,
  conceptIds = NULL,
  CalendarMonths = NULL,
  CalendarYears = NULL
)
}
\arguments{
\item{dataSource}{A list object that is the output of \code{createDatabaseDataSource} or \code{createFileDataSource} function. 
This object helps direct the function to query data from the database (created by \code{createDatabaseDataSource})
or a local premerged file (created by \code{createFileDataSource}). Premerged files are output of cohortDiagnostics
compiled into RData using \code{preMergeDiagnosticsFiles}. Database DataSources are data inserted into a remote
database (only a postgres database is supported) with tables created with DDL function \code{createResultsDataModel}
and uploaded using \code{uploadResults}}

\item{databaseIds}{A vector one or more databaseIds to retrieve the results for. This is a 
character field values from the `databaseId` field of the `database` table 
of the results data model.}

\item{conceptIds}{A vector of one or more Concept Ids.}

\item{calendarMonth}{A vector of one more more calendar month (valid values 0 to 12) where 1 to 12 represents calendar month, and 0 represents
aggregation across all months.}

\item{calendarYears}{A vector of one more more calendar years (valid values 1980 to current year or 0). Where 0 represents
aggregation across all available years.}
}
\value{
Returns a data frame (tibble)
}
\description{
Returns data from concept_count table of Cohort Diagnostics results data model
}
