% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shared.R
\name{getResultsConceptCountSummary}
\alias{getResultsConceptCountSummary}
\title{Returns summary data from concept_count table of Cohort Diagnostics results data model}
\usage{
getResultsConceptCountSummary(
  dataSource,
  databaseIds,
  conceptIds,
  minDate = NULL,
  maxDate = NULL
)
}
\arguments{
\item{dataSource}{A list object that is the output of \code{createDatabaseDataSource} or \code{createFileDataSource} function. 
This object helps direct the function to query data from the database (created by \code{createDatabaseDataSource})
or a local premerged file (created by \code{createFileDataSource}). Premerged files are output of cohortDiagnostics
compiled into RData using \code{preMergeDiagnosticsFiles}. Database DataSources are data inserted into a remote
database (only a postgres database is supported) with tables created with DDL function \code{createResultsDataModel}
and uploaded using \code{uploadResults}}

\item{databaseIds}{A vector one or more databaseIds to retrieve the results for. This is a 
character field values from the `databaseId` field of the `database` table 
of the results data model.}

\item{conceptIds}{A list of concept ids to get counts for}

\item{minDate}{Minimum date of range}

\item{maxDate}{Maximum date of range}
}
\value{
Returns a data frame (tibble)
}
\description{
Returns summary data from concept_count table of Cohort Diagnostics results data model
}
