% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shared.R
\name{getResultsConceptSet}
\alias{getResultsConceptSet}
\title{Returns data from Concept sets table of Cohort Diagnostics results data model}
\usage{
getResultsConceptSet(dataSource, cohortIds = NULL, conceptSetIds = NULL)
}
\arguments{
\item{dataSource}{A list object that is the output of \code{createDatabaseDataSource} or \code{createFileDataSource} function. 
This object helps direct the function to query data from the database (created by \code{createDatabaseDataSource})
or a local premerged file (created by \code{createFileDataSource}). Premerged files are output of cohortDiagnostics
compiled into RData using \code{preMergeDiagnosticsFiles}. Database DataSources are data inserted into a remote
database (only a postgres database is supported) with tables created with DDL function \code{createResultsDataModel}
and uploaded using \code{uploadResults}}

\item{cohortIds}{A vector of one or more Cohort Ids.}

\item{conceptSetIds}{A vector of one or more Concept set Ids.}
}
\value{
Returns a data frame (tibble)
}
\description{
Returns data from Concept sets table of Cohort Diagnostics results data model
}
