% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shared.R
\name{getResultsConceptSetExpression}
\alias{getResultsConceptSetExpression}
\title{Returns a Concept Set expression (R-object) from a cohort definition}
\usage{
getResultsConceptSetExpression(dataSource, cohortId, conceptSetId)
}
\arguments{
\item{dataSource}{A list object that is the output of \code{createDatabaseDataSource} or \code{createFileDataSource} function. 
This object helps direct the function to query data from the database (created by \code{createDatabaseDataSource})
or a local premerged file (created by \code{createFileDataSource}). Premerged files are output of cohortDiagnostics
compiled into RData using \code{preMergeDiagnosticsFiles}. Database DataSources are data inserted into a remote
database (only a postgres database is supported) with tables created with DDL function \code{createResultsDataModel}
and uploaded using \code{uploadResults}}

\item{cohortId}{A single integer value representing the cohort id.}

\item{conceptSetId}{A single integer value representing the concept set id in a cohort definition.}
}
\value{
Returns a data frame (tibble)
}
\description{
Returns a Concept Set expression from a cohort definition
}
