% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsDataModel.R
\name{getResultsDataModelSpecification}
\alias{getResultsDataModelSpecification}
\title{Get specification for Cohort Diagnostics results data model}
\usage{
getResultsDataModelSpecification()
}
\value{
A tibble data frame object with specifications
}
\description{
Get specification for Cohort Diagnostics results data model
}
\examples{
\dontrun{
resultsDataModelSpecification <- getResultsDataModelSpecification()
}

}
