% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsDataModel.R
\name{getResultsDataModelSpecifications}
\alias{getResultsDataModelSpecifications}
\title{Get specifications for Cohort Diagnostics results data model}
\usage{
getResultsDataModelSpecifications(tableName = NULL)
}
\arguments{
\item{tableName}{The name of the table in the results data model to return specification for. 
If NULL (default) then all tables in the results data model will be returned.}
}
\value{
A tibble data frame object with specifications
}
\description{
Get specifications for Cohort Diagnostics results data model
}
