% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsDataModelQueries.R
\name{getResultsFromCohortCount}
\alias{getResultsFromCohortCount}
\title{Returns data from cohort_count table in cohort diagnostics}
\usage{
getResultsFromCohortCount(dataSource, cohortIds, databaseIds)
}
\arguments{
\item{dataSource}{A list object that is the output of \code{createDatabaseDataSource} function.}

\item{cohortIds}{A vector of one or more Cohort Ids.}

\item{databaseIds}{A vector one or more databaseIds to retrieve the results for. This is a 
character field values from the `databaseId` field of the `database` table 
of the results data model.}
}
\value{
Returns a data frame (tibble) with results that conform to cohort counts
table in Cohort Diagnostics results data model.
}
\description{
Returns data from cohort_count table in cohort diagnostics
}
