% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsDataModelQueries.R
\name{getResultsFromOrphanConcept}
\alias{getResultsFromOrphanConcept}
\title{Returns data from orphan_concept table of Cohort Diagnostics results data model}
\usage{
getResultsFromOrphanConcept(dataSource, cohortIds, databaseIds)
}
\arguments{
\item{dataSource}{A list object that is the output of \code{createDatabaseDataSource} function.}

\item{cohortIds}{A vector of one or more Cohort Ids.}

\item{databaseIds}{A vector one or more databaseIds to retrieve the results for. This is a 
character field values from the `databaseId` field of the `database` table 
of the results data model.}
}
\value{
Returns a data frame (tibble) with results that conform to orphan_concept
table in Cohort Diagnostics results data model.
}
\description{
Returns data from orphan_concept table of Cohort Diagnostics results data model
}
