% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsDataModelQueries.R
\name{getResultsIncidenceRateFromPremergedFile}
\alias{getResultsIncidenceRateFromPremergedFile}
\title{Get incidence rates from premerged file with data in results data model format}
\usage{
getResultsIncidenceRateFromPremergedFile(premergedFile, cohortIds, databaseIds)
}
\arguments{
\item{premergedFile}{The full path to read the premerged file.}

\item{cohortIds}{One or more cohort ids to returns incidence rates}

\item{databaseIds}{One or more database ids to returns incidence rates}
}
\value{
Returns a tibble dataframe
}
\description{
Given a set of one or more cohortIds and one or more databaseIds,
return data from the cohort_count table in results data model
with data filtered to the cohortIds and databaseIds.
}
