% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsDataModelQueries.R
\name{getResultsInclusionRuleFromPremergedFile}
\alias{getResultsInclusionRuleFromPremergedFile}
\title{Get inclusion rules from premerged file with data in results data model format}
\usage{
getResultsInclusionRuleFromPremergedFile(premergedFile, cohortIds, databaseIds)
}
\arguments{
\item{premergedFile}{The full path to read the premerged file.}

\item{cohortIds}{One or more cohort ids to returns inclusion rules}

\item{databaseIds}{One or more database ids to returns inclusion rules}
}
\value{
Returns a tibble dataframe
}
\description{
Given a set of one or more cohortIds and one or more databaseIds,
return data from the cohort_count table in results data model
with data filtered to the cohortIds and databaseIds.
}
