% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shared.R
\name{getResultsResolveMappedConceptSet}
\alias{getResultsResolveMappedConceptSet}
\title{Returns resolved and mapped concepts for concept set expression in a cohort}
\usage{
getResultsResolveMappedConceptSet(
  dataSource,
  databaseIds = NULL,
  cohortIds = NULL
)
}
\arguments{
\item{dataSource}{A list object that is the output of \code{createDatabaseDataSource} or \code{createFileDataSource} function. 
This object helps direct the function to query data from the database (created by \code{createDatabaseDataSource})
or a local premerged file (created by \code{createFileDataSource}). Premerged files are output of cohortDiagnostics
compiled into RData using \code{preMergeDiagnosticsFiles}. Database DataSources are data inserted into a remote
database (only a postgres database is supported) with tables created with DDL function \code{createResultsDataModel}
and uploaded using \code{uploadResults}}

\item{databaseIds}{A vector one or more databaseIds to retrieve the results for. This is a 
character field values from the `databaseId` field of the `database` table 
of the results data model.}

\item{cohortIds}{A vector of one or more Cohort Ids.}
}
\value{
Returns a list object with resolved and mapped concepts for all concept sets in one
cohort from one or more data sources. There will be two objects in the returned
list object resolved, mapped - each will be tibble.
}
\description{
Returns a list object with resolved and mapped concepts for all concept sets in one
cohort from one or more data sources. This is being returned from the results data model
of Cohort Diagnostics and is precomputed.
}
