% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsDataModelQueries.R
\name{getResultsTimeDistributionFromCsv}
\alias{getResultsTimeDistributionFromCsv}
\title{Get time distributions from csv file with data in results data model format}
\usage{
getResultsTimeDistributionFromCsv(csvFile, cohortIds, databaseIds)
}
\arguments{
\item{csvFile}{The full path to read the csv file.}

\item{cohortIds}{One or more cohort ids to returns time distributions}

\item{databaseIds}{One or more database ids to returns time distributions}
}
\value{
Returns a tibble dataframe
}
\description{
Given a set of one or more cohortIds and one or more databaseIds,
return data from the cohort_count table in results data model
with data filtered to the cohortIds and databaseIds.
}
