% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsDataModelQueries.R
\name{getResultsTimeSeriesFromDatabase}
\alias{getResultsTimeSeriesFromDatabase}
\title{Get cohort counts from Database with data in results data model format}
\usage{
getResultsTimeSeriesFromDatabase(
  connectionDetails = NULL,
  connection = NULL,
  cohortIds,
  databaseIds
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{cohortIds}{One or more cohort ids to returns cohort counts}

\item{databaseIds}{One or more database ids to returns cohort counts}
}
\value{
Returns a tibble dataframe
}
\description{
Given a set of one or more cohortIds and one or more databaseIds,
return data from the cohort_count table in results data model
with data filtered to the cohortIds and databaseIds.
}
