% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortStats.R
\name{getStatsTable}
\alias{getStatsTable}
\title{Get stats data}
\usage{
getStatsTable(
  connection,
  cohortDatabaseSchema,
  table,
  snakeCaseToCamelCase = FALSE,
  databaseId = NULL,
  includeDatabaseId = TRUE
)
}
\arguments{
\item{connection}{db connection}

\item{cohortDatabaseSchema}{cohort db schema}

\item{table}{table name}

\item{snakeCaseToCamelCase}{snake case to camel case?}

\item{databaseId}{database identifier}

\item{includeDatabaseId}{if db id should be included}
}
\value{
the stats table
}
\description{
Get stats data
}
