% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsPlot.R
\name{getTimeDistributionPlot}
\alias{getTimeDistributionPlot}
\title{Get ggplot object with time distribution plot.}
\usage{
getTimeDistributionPlot(data)
}
\arguments{
\item{data}{A tibble data frame object that is the output of \code{\link{getTimeDistribution}} function.}
}
\value{
A ggplot object.
}
\description{
Get ggplot object with time distribution plot.
}
\examples{
\dontrun{
timeDistributionPlot <- getTimeDistributionPlot(data = data)
}

}
