% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConceptIds.R
\name{getUniqueConceptIds}
\alias{getUniqueConceptIds}
\title{Get all unique concept id's referenced in the cohort diagnostics}
\usage{
getUniqueConceptIds(exportFolder)
}
\arguments{
\item{exportFolder}{The folder where the output is exported by Cohort Diagnostics.
If this folder does not exist, or does not have the searched file
the function will return an error.}
}
\value{
Returns a vector unique conceptId's from various objects in the export folder.
}
\description{
Get all unique concept id's referenced in the cohort diagnostics
}
