% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsDataModel.R
\name{guessCsvFileSpecification}
\alias{guessCsvFileSpecification}
\title{Guesses data model specification from multiple csv files.}
\usage{
guessCsvFileSpecification(pathToCsvFile)
}
\arguments{
\item{pathToCsvFile}{file system path to csv file}
}
\value{
A tibble data frame object with specifications

#' @examples
\dontrun{
csvFileSpecification <- guessCsvFileSpecification(path)
}
}
\description{
Guesses data model specification from multiple csv files.
}
