% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsDataModel.R
\name{guessModelSpecificationForCsv}
\alias{guessModelSpecificationForCsv}
\title{Guesses data model specification for a csv file and makes its available for inspection.}
\usage{
guessModelSpecificationForCsv(pathToCsvFile)
}
\arguments{
\item{pathToCsvFile}{file path to csv file}
}
\description{
Guesses data model specification for a csv file and makes its available for inspection.
}
