% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadDatabaseFiles.R
\name{importCsvFilesToPostgres}
\alias{importCsvFilesToPostgres}
\title{Import csv files in to postgres database instance}
\usage{
importCsvFilesToPostgres(
  connectionDetails,
  schemaName,
  pathToCsvFiles,
  winPsqlPath = ""
)
}
\arguments{
\item{connectionDetails}{DatabaseConnector connection details compatible object. 
Must be dbms="postgres"}

\item{schemaName}{Path to cohort diagnostics csv files to export to}

\item{pathToCsvFiles}{Path to cohort diagnostics (results) csv files to import from}

\item{winPsqlPath}{Path to folder containing postgres executables 
e.g. "C:/Program Files/PostgresSQL/12/bin"}
}
\description{
This utility is a wrapper around the psql command to 
allow uploading large datasets to a postgres database
}
