% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiagnosticInitialization.R
\name{initializeDiagnostics}
\alias{initializeDiagnostics}
\title{Initialize diagnostics}
\usage{
initializeDiagnostics(context, cohortDefinitionSet, cohortIds = NULL)
}
\arguments{
\item{context}{A DiagnosticsContext object.}

\item{cohortDefinitionSet}{Data.frame of cohorts must include columns cohortId, cohortName, json, sql}

\item{cohortIds}{Optionally, provide a subset of cohort IDs to restrict the diagnostics to.}
}
\value{
The modified DiagnosticsContext object.
}
\description{
Initializes shared resources and validates inputs for cohort diagnostics.
}
