% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortConstruction.R
\name{instantiateCohortSet}
\alias{instantiateCohortSet}
\title{Instantiate a set of cohort}
\usage{
instantiateCohortSet(
  connectionDetails = NULL,
  connection = NULL,
  cdmDatabaseSchema,
  oracleTempSchema = NULL,
  cohortDatabaseSchema = cdmDatabaseSchema,
  cohortTable = "cohort",
  baseUrl,
  cohortSetReference,
  generateInclusionStats = FALSE,
  inclusionStatisticsFolder = NULL
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{oracleTempSchema}{Should be used in Oracle to specify a schema where the user has write
privileges for storing temporary tables.}

\item{cohortDatabaseSchema}{Schema name where your cohort table resides. Note that for SQL Server,
this should include both the database and schema name, for example
'scratch.dbo'.}

\item{cohortTable}{Name of the cohort table.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}

\item{cohortSetReference}{A data frame with four columns, as described in the details.}

\item{generateInclusionStats}{Compute and store inclusion rule statistics?}

\item{inclusionStatisticsFolder}{The folder where the inclusion rule statistics are stored. Can be
left NULL if \code{generateInclusionStats = FALSE}.}
}
\description{
This function instantiates a set of cohort in the cohort table, using definitions that are fetched from a WebApi interface.
Optionally, the inclusion rule statistics are computed and stored in the \code{inclusionStatisticsFolder}.
}
\details{
The \code{cohortSetReference} argument must be a data frame with  the following columns:
\describe{
\item{atlasId}{The cohort ID in ATLAS.}
\item{atlasName}{The full name of the cohort. This will be shown in the Shiny app.}
\item{cohortId}{The cohort ID to use in the package. USually the same as the cohort ID in ATLAS.}
\item{name}{A short name for the cohort, to use to create file names. do not use special characters.}
}
}
