% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseFunctions.R
\name{isCDMConnection}
\alias{isCDMConnection}
\title{isCDMConnection}
\usage{
isCDMConnection(connection)
}
\arguments{
\item{connection}{db connection}
}
\value{
if this is CDM connection
}
\description{
isCDMConnection
}
