% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shiny.R
\name{launchDiagnosticsExplorer}
\alias{launchDiagnosticsExplorer}
\title{Launch the Diagnostics Explorer Shiny app}
\usage{
launchDiagnosticsExplorer(
  dataFolder,
  runOverNetwork = FALSE,
  port = 80,
  launch.browser = FALSE
)
}
\arguments{
\item{dataFolder}{A folder where the exported zip files for the diagnostics are stored. Use
the \code{\link{runCohortDiagnostics}} function to generate these zip files. 
Zip files containing results from multiple databases can be placed in the same
folder.}

\item{runOverNetwork}{(optional) Do you want the app to run over your network?}

\item{port}{(optional) Only used if \code{runOverNetwork} = TRUE.}

\item{launch.browser}{Should the app be launched in your default browser, or in a Shiny window.
Note: copying to clipboard will not work in a Shiny window.}
}
\description{
Launch the Diagnostics Explorer Shiny app
}
\details{
Launches a Shiny app that allows the user to explore the diagnostics
}
