% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortConstruction.R
\name{loadCohortsFromPackage}
\alias{loadCohortsFromPackage}
\title{Load Cohort Definitions From A Study Package}
\usage{
loadCohortsFromPackage(
  packageName,
  cohortToCreateFile = "settings/cohortsToCreate.csv",
  cohortIds = NULL,
  errorMessage = NULL
)
}
\arguments{
\item{packageName}{The name of the package containing the cohort definitions. Can be left NULL if 
\code{baseUrl} and \code{cohortSetReference} have been specified.}

\item{cohortToCreateFile}{The location of the cohortToCreate file within the package. Is ignored if 
\code{baseUrl} and \code{cohortSetReference} have been specified. The
cohortToCreateFile must be .csv file that is expected to be read into 
a dataframe object identical to requirements for \code{cohortSetReference} argument. 
This csv file is expected to be encoded in either ASCII or UTF-8, if not, an 
error message will be displayed and process stopped.}

\item{cohortIds}{Optionally, provide a subset of cohort IDs to restrict the
diagnostics to.}

\item{errorMessage}{checkmate assert collection, used internally for error checks}
}
\description{
Load cohort references for usage in executeDiagnostics.

Expects a csv file (e.g one placed in "inst/settings/cohortsToCreate.csv" for the pacakge) with headers:
atlasId, referentConceptId, webApiCohortId, cohortId, name,
cohortName, logicDescription

For example
```
atlasId,referentConceptId,webApiCohortId,cohortId,name,cohortName,logicDescription,phenotypeId
17492,192671,17492,17492,17492,GI bleed,GI bleed,192671000
```
}
