% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsPlot.R
\name{plotCohortCompare}
\alias{plotCohortCompare}
\title{Get Plotly object with cohort comparison plot.}
\usage{
plotCohortCompare(balance, cohortId, comparatorId)
}
\arguments{
\item{balance}{A tibble data frame object that is the output of \code{\link{getCohortCompare}} function.
cohortId   A input value given, when the user select the cohort in the shiny app.
comparatorId   A input value given, when the user select the comparator in the shiny app.}
}
\value{
A Plotly object.
}
\description{
Get Plotly object with cohort comparison plot.
}
\examples{
\dontrun{
plotCohortCompare <- getCohortCompare(data = data)
}

}
