% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsPlot.R
\name{plotCohortOverlap}
\alias{plotCohortOverlap}
\title{Get Vendiagram object with cohort Overlap plot.}
\usage{
plotCohortOverlap(data)
}
\arguments{
\item{data}{A tibble data frame object that is the output of \code{\link{getCohortOverlap}} function.}
}
\value{
A Vendiagram object.
}
\description{
Get Vendiagram  object with cohort Overlap plot.
}
\examples{
\dontrun{
plotCohortOverlap <- getCohortOverlap(data = data)
}

}
