% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsPlot.R
\name{plotCohortOverlapVennDiagram}
\alias{plotCohortOverlapVennDiagram}
\title{Get Vendiagram object with cohort Overlap plot.}
\usage{
plotCohortOverlapVennDiagram(
  data,
  targetCohortIds = NULL,
  comparatorCohortIds = NULL,
  databaseIds = NULL
)
}
\arguments{
\item{data}{A tibble data frame object that is the output of 
\code{\link{getCohortOverlap}} function.}

\item{targetCohortIds}{(optional) A vector of one or more Cohort Ids.}

\item{comparatorCohortIds}{(optional) A vector of one or more Cohort Ids.}

\item{databaseIds}{A vector one or more databaseIds to retrieve the results for. This is a 
character field values from the `databaseId` field of the `database` table 
of the results data model.}
}
\value{
A Vendiagram object.
}
\description{
Get Vendiagram  object with cohort Overlap plot.
}
\examples{
\dontrun{
plotCohortOverlapVennDiagram <- plotCohortOverlapVennDiagram(data = data)
}

}
