% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsPlot.R
\name{plotCohortStandardizedDifferenceComparison}
\alias{plotCohortStandardizedDifferenceComparison}
\title{Get Plotly object with cohort comparison plot.}
\usage{
plotCohortStandardizedDifferenceComparison(
  data,
  targetCohortIds = NULL,
  comparatorCohortIds = NULL,
  cohort = NULL,
  covariateReference = NULL,
  concept = NULL,
  databaseIds = NULL
)
}
\arguments{
\item{targetCohortIds}{A vector of one or more Cohort Ids.}

\item{comparatorCohortIds}{A vector of one or more Cohort Ids.}

\item{covariateReference}{(optional) A tibble data frame object returned from \code{\link{getCovariateReference}} function.}

\item{concept}{(optional) A tibble data frame object returned from \code{\link{getConceptReference}} function.}

\item{databaseIds}{A vector one or more databaseIds to retrieve the results for. This is a 
character field values from the `databaseId` field of the `database` table 
of the results data model.}

\item{A}{tibble data frame object that is the output of \code{\link{compareCovariateValueResult}} function.}

\item{cohorts}{(optional) A tibble data frame object returned from \code{\link{getCohortReference}} function.}
}
\value{
A Plotly object.
}
\description{
Get Plotly object with cohort comparison plot.
}
\examples{
\dontrun{
plotCohortCompare <- getCohortCompare(data = data)
}

}
