% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsPlot.R
\name{plotIncidenceRate}
\alias{plotIncidenceRate}
\title{Get ggplot object with incidence rate plot.}
\usage{
plotIncidenceRate(
  data,
  stratifyByAge = TRUE,
  stratifyByGender = TRUE,
  stratifyByCalendarYear = TRUE,
  yscaleFixed = FALSE
)
}
\arguments{
\item{data}{A tibble data frame object that is the output of \code{\link{getIncidenceRate}} function.}

\item{stratifyByAge}{Do you want to stratify by age?}

\item{stratifyByGender}{Do you want to stratify by gender?}

\item{stratifyByCalendarYear}{Do you want to stratify by calendar year?}

\item{yscaleFixed}{Do you want to rescale y-axis?}
}
\value{
A ggplot object.
}
\description{
Get ggplot object with incidence rate plot.
}
\examples{
\dontrun{
timeDistributionPlot <- getTimeDistributionPlot(data = data)
}

}
