% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLogFile.R
\name{plotLogFile}
\alias{plotLogFile}
\title{plotLogFile}
\usage{
plotLogFile(logPath)
}
\arguments{
\item{logPath}{Path to the generated logfile}
}
\value{
`ggplot`
}
\description{
plotLogFile
}
\examples{
logPath <- "./path/to/log.txt"
if (file.exists(logPath)) {
  plotLogFile(logPath)
}
}
