% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsDataModel.R
\name{postProcessDiagnosticsResultsFiles}
\alias{postProcessDiagnosticsResultsFiles}
\title{Prepare diagnostics results for use}
\usage{
postProcessDiagnosticsResultsFiles(dataFolder, outputFolder)
}
\arguments{
\item{dataFolder}{folder where the exported zip files for the diagnostics are stored. Use
the \code{\link{runCohortDiagnostics}} function to generate these zip files. 
Zip files containing results from multiple databases may be placed in the same
folder.}

\item{outputFolder}{folder where the post processed files for the diagnostics are to be stored. 
These files may be used with the results viewer or may be uploaded into RDMS.}
}
\description{
This function takes as input one or more .zip files that are the output of
\code{\link{runCohortDiagnostics}} function, performs a set of data management and 
data quality checks. Data Management steps involve appending and depulicating data
sourced from different data sources (each should be a seperate zip file). Data
Quality checks involves checks for conformance to name/data type/data constraints
requirements for the Cohort Diagnostics Results data model. The function will 
output two files a 
- zip file: a compressed zip file with csv that is ready for upload into a rdms 
system with the schema/table/constraints compatible with the Cohort Diagnostics
results data model (see '/settings/resultsDataModelSpecification.csv')
- rds file: called preMerged.RData that may used in the data folder of a Shiny 
app. This rds data when placed in the 'data' folder of the shiny app will 
be automatically recognized by the Shiny app.

Note: the csv files in the .zip files are expected to have a snake case file
naming convention and snake case column names.
}
