% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseFunctions.R
\name{querySql}
\alias{querySql}
\title{querySql}
\usage{
querySql(
  connection,
  sql,
  errorReportFile = file.path(getwd(), "errorReportSql.txt"),
  snakeCaseToCamelCase = FALSE,
  integerAsNumeric = getOption("databaseConnectorIntegerAsNumeric", default = TRUE),
  integer64AsNumeric = getOption("databaseConnectorInteger64AsNumeric", default = TRUE)
)
}
\arguments{
\item{connection}{db connection}

\item{sql}{The SQL to be executed}

\item{errorReportFile}{error file}

\item{snakeCaseToCamelCase}{snake case to camel case boolean}

\item{integerAsNumeric}{int as numeric boolean}

\item{integer64AsNumeric}{int64 as numeric boolean}
}
\value{
query results
}
\description{
querySql
}
