% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortDiagnostics.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{getCohortCounts}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{CohortGenerator}{\code{\link[CohortGenerator]{getCohortCounts}}}
}}

