% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseFunctions.R
\name{renderTranslateQuerySql}
\alias{renderTranslateQuerySql}
\title{renderTranslateQuerySql}
\usage{
renderTranslateQuerySql(
  connection,
  sql,
  errorReportFile = file.path(getwd(), "errorReportSql.txt"),
  snakeCaseToCamelCase = FALSE,
  oracleTempSchema = NULL,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  integerAsNumeric = getOption("databaseConnectorIntegerAsNumeric", default = TRUE),
  integer64AsNumeric = getOption("databaseConnectorInteger64AsNumeric", default = TRUE),
  ...
)
}
\arguments{
\item{connection}{}

\item{sql}{}

\item{errorReportFile}{}

\item{snakeCaseToCamelCase}{}

\item{oracleTempSchema}{}

\item{tempEmulationSchema}{}

\item{integerAsNumeric}{}

\item{integer64AsNumeric}{}

\item{...}{}
}
\value{
the query results
}
\description{
renderTranslateQuerySql
}
