% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseFunctions.R
\name{renderTranslateQuerySqlToAndromeda}
\alias{renderTranslateQuerySqlToAndromeda}
\title{renderTranslateQuerySqlToAndromeda}
\usage{
renderTranslateQuerySqlToAndromeda(
  connection,
  sql,
  andromeda,
  andromedaTableName,
  errorReportFile = file.path(getwd(), "errorReportSql.txt"),
  snakeCaseToCamelCase = FALSE,
  appendToTable = FALSE,
  integerAsNumeric = getOption("databaseConnectorIntegerAsNumeric", default = TRUE),
  integer64AsNumeric = getOption("databaseConnectorInteger64AsNumeric", default = TRUE),
  ...
)
}
\arguments{
\item{connection}{db connection}

\item{sql}{The SQL to be executed}

\item{andromeda}{andromeda object}

\item{andromedaTableName}{table name}

\item{errorReportFile}{error file}

\item{snakeCaseToCamelCase}{snake case to camel case boolean}

\item{appendToTable}{if result should be appended}

\item{integerAsNumeric}{int as numeric boolean}

\item{integer64AsNumeric}{int64 as numeric boolean}

\item{...}{parameters that will be used to render the SQL}
}
\value{
the updated andromeda object
}
\description{
renderTranslateQuerySqlToAndromeda
}
