% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCohortCharacterization.R
\name{runCohortCharacterization}
\alias{runCohortCharacterization}
\title{runCohortCharacterization}
\usage{
runCohortCharacterization(
  connection,
  databaseId,
  exportFolder,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTable,
  covariateSettings,
  tempEmulationSchema,
  cdmVersion,
  cohorts,
  cohortCounts,
  minCellCount,
  instantiatedCohorts,
  incremental,
  recordKeepingFile,
  covariateValueFileName = file.path(exportFolder, "temporal_covariate_value.csv"),
  covariateValueContFileName = file.path(exportFolder,
    "temporal_covariate_value_dist.csv"),
  covariateRefFileName = file.path(exportFolder, "temporal_covariate_ref.csv"),
  analysisRefFileName = file.path(exportFolder, "temporal_analysis_ref.csv"),
  timeRefFileName = file.path(exportFolder, "temporal_time_ref.csv"),
  minCharacterizationMean = 0.001,
  batchSize = getOption("CohortDiagnostics-FE-batch-size", default = 20)
)
}
\arguments{
\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package.}

\item{databaseId}{A short string for identifying the database (e.g. 'Synpuf').}

\item{exportFolder}{The folder where the results will be exported to}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{cohortDatabaseSchema}{Schema name where your cohort table resides. Note that for SQL Server,
this should include both the database and schema name, for example
'scratch.dbo'.}

\item{cohortTable}{Name of the cohort table.}

\item{covariateSettings}{Either an object of type \code{covariateSettings} as created using one of
the createTemporalCovariateSettings function in the FeatureExtraction package, or a list
of such objects.}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not truly support temp tables. To emulate temp 
tables, provide a schema with write privileges where temp tables can be created.}

\item{cdmVersion}{The version of the OMOP CDM. Default 5. (Note: only 5 is supported.)}

\item{cohorts}{The cohorts for which the covariates need to be obtained}

\item{cohortCounts}{A dataframe with the cohort counts}

\item{minCellCount}{The minimum cell count for fields contains person counts or fractions}

\item{instantiatedCohorts}{cohortIds of the cohorts that have been already been instantiated}

\item{incremental}{`TRUE` or `FALSE` (default). If TRUE diagnostics for cohorts in the 
cohort definition set that have not changed will be skipped and existing results 
csv files will be updated. If FALSE then diagnostics for all cohorts in the cohort 
definition set will be executed and pre-existing results files will be deleted.}

\item{covariateValueFileName}{Filename of the binary covariates output}

\item{covariateValueContFileName}{Filename of the continuous covariate output}

\item{covariateRefFileName}{Filename of the covariate reference output}

\item{analysisRefFileName}{Filename of the analysis reference output}

\item{timeRefFileName}{Filename of the time reference output}

\item{minCharacterizationMean}{The minimum mean value for characterization output. Values below this will be cut off from output. This
will help reduce the file size of the characterization output, but will remove information
on covariates that have very low values. The default is 0.001 (i.e. 0.1 percent)}

\item{batchSize}{In case of batch processing, this specifies the size of the batch}

\item{incrementalFolder}{If \code{incremental = TRUE}, specify a folder where records are kept
of which cohort diagnostics has been executed.}
}
\value{
None, it will write results to disk
}
\description{
This function takes cohorts as input and generates the covariates for these cohorts.
The covariates are generated using FeatureExtraction. The output from this package 
is slightly modified before the output is written to disk. 
These are the files written to disk, if available:
 * temporal_analysis_ref.csv
 * temporal_covariate_ref.csv
 * temporal_covariate_value.csv
 * temporal_covariate_value_dist.csv
 * temporal_time_ref.csv
}
