% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunDiagnosticsExternalCounts.R
\name{runCohortDiagnosticsExternalCounts}
\alias{runCohortDiagnosticsExternalCounts}
\title{Run cohort diagnostics using external concept counts}
\usage{
runCohortDiagnosticsExternalCounts(
  packageName,
  cohortToCreateFile = "settings/CohortsToCreate.csv",
  connectionDetails = NULL,
  connection = NULL,
  cdmDatabaseSchema,
  oracleTempSchema = NULL,
  cohortDatabaseSchema,
  cohortTable = "cohort",
  cohortIds = NULL,
  conceptCountsDatabaseSchema = cdmDatabaseSchema,
  conceptCountsTable = "concept_counts",
  conceptCountsTableIsTemp = FALSE,
  exportFolder,
  databaseId,
  databaseName,
  databaseDescription,
  runIncludedSourceConcepts = TRUE,
  runOrphanConcepts = TRUE,
  minCellCount = 5
)
}
\arguments{
\item{packageName}{The name of the package containing the cohort definitions}

\item{cohortToCreateFile}{The location of the cohortToCreate file within the package.}

\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{oracleTempSchema}{Should be used in Oracle to specify a schema where the user has write
privileges for storing temporary tables.}

\item{cohortDatabaseSchema}{Schema name where your cohort table resides. Note that for SQL Server,
this should include both the database and schema name, for example
'scratch.dbo'.}

\item{cohortTable}{Name of the cohort table.}

\item{cohortIds}{Optionally, provide a subset of cohort IDs to restrict the
diagnostics to.}

\item{conceptCountsDatabaseSchema}{Schema name where your concept counts table resides. Note that
for SQL Server, this should include both the database and
schema name, for example 'scratch.dbo'. Ignored if
\code{conceptCountsTableIsTemp = TRUE}.}

\item{conceptCountsTable}{Name of the concept counts table. This table can be created
using the \code{\link{createConceptCountsTable}}.}

\item{conceptCountsTableIsTemp}{Is the concept counts table a temp table?}

\item{exportFolder}{The folder where the output will be exported to. If this folder
does not exist it will be created.}

\item{databaseId}{A short string for identifying the database (e.g. 'Synpuf').}

\item{databaseName}{The full name of the database.}

\item{databaseDescription}{A short description (several sentences) of the database.}

\item{runIncludedSourceConcepts}{Generate and export the source concepts included in the cohorts?}

\item{runOrphanConcepts}{Generate and export potential orphan concepts?}

\item{minCellCount}{The minimum cell count for fields contains person counts or fractions.}
}
\description{
Runs the cohort diagnostics on all (or a subset of) the cohorts instantiated using the
\code{ROhdsiWebApi::insertCohortDefinitionSetInPackage} function, but using external concept counts.
}
