% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunDiagnosticsExternalCounts.R
\name{runCohortDiagnosticsUsingExternalCounts}
\alias{runCohortDiagnosticsUsingExternalCounts}
\title{Run cohort diagnostics using external concept counts}
\usage{
runCohortDiagnosticsUsingExternalCounts(
  packageName = NULL,
  cohortToCreateFile = "settings/CohortsToCreate.csv",
  baseUrl = NULL,
  cohortSetReference = NULL,
  connectionDetails = NULL,
  connection = NULL,
  cdmDatabaseSchema,
  oracleTempSchema = NULL,
  cohortIds = NULL,
  conceptCountsDatabaseSchema = cdmDatabaseSchema,
  conceptCountsTable = "concept_counts",
  conceptCountsTableIsTemp = FALSE,
  exportFolder,
  databaseId,
  databaseName = databaseId,
  databaseDescription = "",
  runIncludedSourceConcepts = TRUE,
  runOrphanConcepts = TRUE,
  minCellCount = 5
)
}
\arguments{
\item{packageName}{The name of the package containing the cohort definitions. Can be left NULL if 
\code{baseUrl} and \code{cohortSetReference} have been specified.}

\item{cohortToCreateFile}{The location of the cohortToCreate file within the package. Is ignored if 
\code{baseUrl} and \code{cohortSetReference} have been specified.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI". Can be left NULL if 
\code{packageName} and \code{cohortToCreateFile} have been specified.}

\item{cohortSetReference}{A data frame with four columns, as described in the details. Can be left NULL if 
\code{packageName} and \code{cohortToCreateFile} have been specified.}

\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{oracleTempSchema}{Should be used in Oracle to specify a schema where the user has write
privileges for storing temporary tables.}

\item{cohortIds}{Optionally, provide a subset of cohort IDs to restrict the
diagnostics to.}

\item{conceptCountsDatabaseSchema}{Schema name where your concept counts table resides. Note that
for SQL Server, this should include both the database and
schema name, for example 'scratch.dbo'. Ignored if
\code{conceptCountsTableIsTemp = TRUE}.}

\item{conceptCountsTable}{Name of the concept counts table. This table can be created
using the \code{\link{createConceptCountsTable}}.}

\item{conceptCountsTableIsTemp}{Is the concept counts table a temp table?}

\item{exportFolder}{The folder where the output will be exported to. If this folder
does not exist it will be created.}

\item{databaseId}{A short string for identifying the database (e.g. 'Synpuf').}

\item{databaseName}{The full name of the database.}

\item{databaseDescription}{A short description (several sentences) of the database.}

\item{runIncludedSourceConcepts}{Generate and export the source concepts included in the cohorts?}

\item{runOrphanConcepts}{Generate and export potential orphan concepts?}

\item{minCellCount}{The minimum cell count for fields contains person counts or fractions.}
}
\description{
Runs  cohort diagnostics on all (or a subset of) the cohorts, but using external concept counts. The external counts 
must have the following columns:

\describe{
\item{concept_id}{The source or target concept ID.}
\item{concept_count}{The number of records having the concept.}
\item{concept_subjects}{The number of unique persons having the concept.}
}
}
\details{
Currently two ways of executing this function are supported, either (1) embedded in a study package, assuming 
the cohort definitions are stored in that package using the \code{ROhdsiWebApi::insertCohortDefinitionSetInPackage},
or (2) by using a WebApi interface to retrieve the cohort definitions.

When using this function from within a study package, use the \code{packageName} and \code{cohortToCreateFile} to specify
the name of the study package, and the name of the cohortToCreate file within that package, respectivly

When using this function using a WebApi interface, use the \code{baseUrl} and \code{cohortSetReference} to specify how to 
connect to the WebApi, and which cohorts to fetch, respectively.

The \code{cohortSetReference} argument must be a data frame with  the following columns:
\describe{
\item{atlasId}{The cohort ID in ATLAS.}
\item{atlasName}{The full name of the cohort. This will be shown in the Shiny app.}
\item{cohortId}{The cohort ID to use in the package. USually the same as the cohort ID in ATLAS.}
\item{name}{A short name for the cohort, to use to create file names. do not use special characters.}
}
}
