% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCohortRelationship.R
\name{runCohortRelationship}
\alias{runCohortRelationship}
\title{Generate and export the cohort temporal relationships}
\usage{
runCohortRelationship(
  connection,
  cohortDefinitionSet,
  exportFolder,
  databaseId,
  cohortDatabaseSchema,
  cdmDatabaseSchema,
  tempEmulationSchema,
  cohortTable,
  temporalCovariateSettings,
  minCellCount,
  incremental,
  incrementalFolder = exportFolder,
  batchSize = getOption("CohortDiagnostics-Relationship-batch-size", default = 500)
)
}
\arguments{
\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package.}

\item{cohortDefinitionSet}{A data.frame with cohort definitions created by 
`CohortGenerator::getCohortDefinitionSet` that must include 
the columns cohortId, cohortName, json, sql.}

\item{exportFolder}{The folder where the results will be exported to}

\item{databaseId}{A short string for identifying the database (e.g. 'Synpuf').}

\item{cohortDatabaseSchema}{Schema name where your cohort table resides. Note that for SQL Server,
this should include both the database and schema name, for example
'scratch.dbo'.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not truly support temp tables. To emulate temp 
tables, provide a schema with write privileges where temp tables can be created.}

\item{cohortTable}{Name of the cohort table.}

\item{temporalCovariateSettings}{Either an object of type covariateSettings as created using one of the createTemporalCovariateSettings function in the FeatureExtraction package, or a list of such objects.}

\item{minCellCount}{The minimum cell count for fields contains person counts or fractions}

\item{incremental}{`TRUE` or `FALSE` (default). If TRUE diagnostics for cohorts in the 
cohort definition set that have not changed will be skipped and existing results 
csv files will be updated. If FALSE then diagnostics for all cohorts in the cohort 
definition set will be executed and pre-existing results files will be deleted.}

\item{incrementalFolder}{If \code{incremental = TRUE}, specify a folder where records are kept
of which cohort diagnostics has been executed.}

\item{batchSize}{In case of batch processing, this specifies the size of the batch}
}
\value{
None, it will write the results to a csv file.
}
\description{
Generate and export the cohort relationship. Cohort relationship checks the temporal relationship
between two or more cohorts and derives subject counts for cohorts with different temporal relationships.
}
