% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConceptSetDiagnostic.R
\name{runConceptSetDiagnostic}
\alias{runConceptSetDiagnostic}
\title{Run concept set diagnostics}
\usage{
runConceptSetDiagnostic(
  context,
  cohortDefinitionSet = NULL,
  cohortIds = NULL,
  runIncludedSourceConcepts = TRUE,
  runOrphanConcepts = TRUE,
  runBreakdownIndexEvents = TRUE
)
}
\arguments{
\item{context}{A DiagnosticsContext object.}

\item{cohortDefinitionSet}{Data.frame of cohorts must include columns cohortId, cohortName, json, sql}

\item{cohortIds}{Optionally, provide a subset of cohort IDs to restrict the diagnostics to.}

\item{runIncludedSourceConcepts}{Generate and export the source concepts included in the cohorts?}

\item{runOrphanConcepts}{Generate and export potential orphan concepts?}

\item{runBreakdownIndexEvents}{Generate and export the breakdown of index events?}
}
\value{
Invisible NULL
}
\description{
Runs the concept set diagnostics (included source concepts, orphan concepts, and/or breakdown of index events).
}
