% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runIncidenceRate.R
\name{runIncidenceRate}
\alias{runIncidenceRate}
\title{Title}
\usage{
runIncidenceRate(
  connection,
  tempEmulationSchema,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTable,
  databaseId,
  exportFolder,
  minCellCount,
  cohorts,
  instantiatedCohorts,
  recordKeepingFile,
  washoutPeriod,
  incremental
)
}
\arguments{
\item{connection}{}

\item{tempEmulationSchema}{}

\item{cdmDatabaseSchema}{}

\item{cohortDatabaseSchema}{}

\item{cohortTable}{}

\item{databaseId}{}

\item{exportFolder}{}

\item{minCellCount}{}

\item{cohorts}{}

\item{instantiatedCohorts}{}

\item{recordKeepingFile}{}

\item{washoutPeriod}{}

\item{incremental}{}
}
\description{
Title
}
