% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IncidenceRateDiagnostic.R
\name{runIncidenceRateDiagnostic}
\alias{runIncidenceRateDiagnostic}
\title{Run incidence rate diagnostic}
\usage{
runIncidenceRateDiagnostic(
  context,
  cohortDefinitionSet = NULL,
  cohortIds = NULL,
  irWashoutPeriod = 0
)
}
\arguments{
\item{context}{A DiagnosticsContext object.}

\item{cohortDefinitionSet}{Data.frame of cohorts must include columns cohortId, cohortName, json, sql}

\item{cohortIds}{Optionally, provide a subset of cohort IDs to restrict the diagnostics to.}

\item{irWashoutPeriod}{Number of days washout to include in calculation of incidence rates - default is 0}
}
\value{
Invisible NULL
}
\description{
Runs the incidence rate diagnostic for a set of cohorts.
}
