% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runIncludedSourceConcepts.R
\name{runIncludedSourceConcepts}
\alias{runIncludedSourceConcepts}
\title{Title}
\usage{
runIncludedSourceConcepts(
  connection,
  cohortDefinitionSet,
  tempEmulationSchema,
  cdmDatabaseSchema,
  databaseId,
  exportFolder,
  minCellCount,
  incremental = FALSE,
  recordKeepingFile
)
}
\arguments{
\item{connection}{}

\item{cohortDefinitionSet}{A data.frame with cohort definitions created by 
`CohortGenerator::getCohortDefinitionSet` that must include 
the columns cohortId, cohortName, json, sql.}

\item{tempEmulationSchema}{}

\item{cdmDatabaseSchema}{}

\item{databaseId}{}

\item{exportFolder}{}

\item{minCellCount}{}

\item{incremental}{}

\item{recordKeepingFile}{}

\item{vocabularyDatabaseSchema}{}

\item{conceptCountsDatabaseSchema}{}

\item{conceptCountsTable}{}

\item{conceptCountsTableIsTemp}{}

\item{cohortDatabaseSchema}{}

\item{cohortTable}{}

\item{useExternalConceptCountsTable}{}

\item{conceptIdTable}{}

\item{resultsDatabaseSchema}{}
}
\description{
Title
}
