% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runResolvedConceptSets.R
\name{runResolvedConceptSets}
\alias{runResolvedConceptSets}
\title{Generates and exports the uniqu conceptSetIds from all cohorts}
\usage{
runResolvedConceptSets(
  connection,
  cohortDefinitionSet,
  databaseId,
  exportFolder,
  minCellCount,
  vocabularyDatabaseSchema,
  tempEmulationSchema,
  incremental = FALSE,
  incrementalFolder = exportFolder
)
}
\arguments{
\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package.}

\item{cohortDefinitionSet}{A data.frame with cohort definitions created by 
`CohortGenerator::getCohortDefinitionSet` that must include 
the columns cohortId, cohortName, json, sql.}

\item{databaseId}{A short string for identifying the database (e.g. 'Synpuf').}

\item{exportFolder}{The folder where the results will be exported to}

\item{minCellCount}{The minimum cell count for fields contains person counts or fractions}

\item{vocabularyDatabaseSchema}{Schema name where your OMOP vocabulary data resides. This is 
commonly the same as cdmDatabaseSchema. Note that for 
SQL Server, this should include both the database and
schema name, for example 'vocabulary.dbo'.}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not truly support temp tables. To emulate temp 
tables, provide a schema with write privileges where temp tables can be created.}

\item{incremental}{`TRUE` or `FALSE` (default). If TRUE diagnostics for cohorts in the 
cohort definition set that have not changed will be skipped and existing results 
csv files will be updated. If FALSE then diagnostics for all cohorts in the cohort 
definition set will be executed and pre-existing results files will be deleted.}

\item{incrementalFolder}{If \code{incremental = TRUE}, specify a folder where records are kept
of which cohort diagnostics has been executed.}
}
\value{
None, it will write the results to a csv file.
}
\description{
Combines the unique conceptSetIds from all cohorts.
}
