% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunDiagnostics.R
\name{runStudyDiagnostics}
\alias{runStudyDiagnostics}
\title{Run study diagnostics}
\usage{
runStudyDiagnostics(packageName,
  cohortToCreateFile = "settings/CohortsToCreate.csv",
  connectionDetails = NULL, connection = NULL, cohortDatabaseSchema,
  cohortTable = "cohort", cohortIds = NULL,
  inclusionStatisticsFolder = NULL, exportFolder,
  runInclusionStatistics = TRUE, runIncludedSourceConcepts = TRUE,
  runOrphanConcepts = TRUE, runBreakdownIndexEvents = TRUE)
}
\arguments{
\item{packageName}{The name of the package containing the cohort definitions}

\item{cohortToCreateFile}{The location of the cohortToCreate file witing the package.}

\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is provided.}

\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails} is provided,
in which case a new connection will be opened at the start of the function, and closed
when the function finishes.}

\item{cohortDatabaseSchema}{Schema name where your cohort table resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'scratch.dbo'.}

\item{cohortTable}{Name of the cohort table.}

\item{cohortIds}{Optionally, provide a subset of cohort IDs to restrict the diagnostics to.}

\item{inclusionStatisticsFolder}{The folder where the inclusion rule statistics are stored. Can be left NULL if 
\code{runInclusionStatistics = FALSE}.}

\item{exportFolder}{The folder where the output will be exported to. If this folder does not exist it will be created.}

\item{runInclusionStatistics}{Generate and export statistic on the cohort incusion rules?}

\item{runIncludedSourceConcepts}{Generate and export the source concepts included in the cohorts?}

\item{runOrphanConcepts}{Generate and export potential orphan concepts?}

\item{runBreakdownIndexEvents}{Generate and export the breakdown of index events?}
}
\description{
Runs the study diagnostics on all (or a subset of) the cohorts instantiated using the \code{ROhdsiWebApi::insertCohortDefinitionSetInPackage}
function.
}
