% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharacterizationDiagnostic.R
\name{runTemporalCharacterizationDiagnostic}
\alias{runTemporalCharacterizationDiagnostic}
\title{Run temporal characterization diagnostic}
\usage{
runTemporalCharacterizationDiagnostic(
  context,
  cohortDefinitionSet = NULL,
  cohortIds = NULL,
  temporalCovariateSettings = getDefaultCovariateSettings(),
  minCharacterizationMean = 0.01,
  runFeatureExtractionOnSample = FALSE,
  sampleN = 1000,
  seed = 64374,
  seedArgs = NULL,
  runCohortRelationship = TRUE
)
}
\arguments{
\item{context}{A DiagnosticsContext object.}

\item{cohortDefinitionSet}{Data.frame of cohorts must include columns cohortId, cohortName, json, sql}

\item{cohortIds}{Optionally, provide a subset of cohort IDs to restrict the diagnostics to.}

\item{temporalCovariateSettings}{Either an object of type \code{covariateSettings} as created using one of
the createTemporalCovariateSettings function in the FeatureExtraction package, or a list
of such objects.}

\item{minCharacterizationMean}{The minimum mean value for characterization output. Values below this will be cut off from output.}

\item{runFeatureExtractionOnSample}{Logical. If TRUE, the function will operate on a sample of the data.}

\item{sampleN}{Integer. The number of records to include in the sample if runFeatureExtractionOnSample is TRUE.}

\item{seed}{Integer. The seed for the random number generator used to create the sample.}

\item{seedArgs}{List. Additional arguments to pass to the sampling function.}

\item{runCohortRelationship}{Compute cohort relationships?}
}
\value{
Invisible NULL
}
\description{
Runs the temporal characterization diagnostic for a set of cohorts.
}
