% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeDynamicNetworkAnalysis.R
\name{runTimeDynamicNetworkAnalysis}
\alias{runTimeDynamicNetworkAnalysis}
\title{Time Dynamic Network Analysis}
\usage{
runTimeDynamicNetworkAnalysis(
  ...,
  timeBinSize = 7,
  timeWindow = c(-365, 365),
  cdSettings = NULL
)
}
\arguments{
\item{...}{Global  cdSettings}

\item{timeBinSize}{The time window to determine that events co-occur}

\item{timeWindow}{vector of length 2, the time in days before and after index to include in analysis}

\item{cdSettings}{CohortDiagnostics settings object}
}
\description{
ALl events that occur in a patients history can be viewed as states in a time dynamic network model.
In this approach we make a series of simplifying assumtions to compute a directed, weighted, time dynamic graph.

Events fall in time windows relative to index date, as specified by the timeBinSize parameter.

For example a condition occurrence 7 days prior to or post index

Nodes represent covariates, edges represent the coccurence at different time points

This network is not exported for individuals, but only in the aggregate with the objective of finding common pathways
prior to index and post index.
}
