% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeriesDiagnostic.R
\name{runTimeSeriesDiagnostic}
\alias{runTimeSeriesDiagnostic}
\title{Run time series diagnostic}
\usage{
runTimeSeriesDiagnostic(context, cohortDefinitionSet = NULL, cohortIds = NULL)
}
\arguments{
\item{context}{A DiagnosticsContext object.}

\item{cohortDefinitionSet}{Data.frame of cohorts must include columns cohortId, cohortName, json, sql}

\item{cohortIds}{Optionally, provide a subset of cohort IDs to restrict the diagnostics to.}
}
\value{
Invisible NULL
}
\description{
Runs the time series diagnostic for a set of cohorts.
}
