% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runVisitContext.R
\name{runVisitContext}
\alias{runVisitContext}
\title{runVisitContext}
\usage{
runVisitContext(
  connection,
  cohortDefinitionSet,
  exportFolder,
  databaseId,
  cohortDatabaseSchema,
  cdmDatabaseSchema,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  cohortTable = "cohort",
  cdmVersion = 5,
  minCellCount,
  incremental,
  incrementalFolder = file.path(exportFolder, "incremental")
)
}
\arguments{
\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package.}

\item{cohortDefinitionSet}{A data.frame with cohort definitions created by 
`CohortGenerator::getCohortDefinitionSet` that must include 
the columns cohortId, cohortName, json, sql.}

\item{exportFolder}{The folder where the results will be exported to}

\item{databaseId}{A short string for identifying the database (e.g. 'Synpuf').}

\item{cohortDatabaseSchema}{Schema name where your cohort table resides. Note that for SQL Server,
this should include both the database and schema name, for example
'scratch.dbo'.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not truly support temp tables. To emulate temp 
tables, provide a schema with write privileges where temp tables can be created.}

\item{cohortTable}{Name of the cohort table.}

\item{cdmVersion}{The version of the OMOP CDM. Default 5. (Note: only 5 is supported.)}

\item{minCellCount}{The minimum cell count for fields contains person counts or fractions}

\item{incremental}{`TRUE` or `FALSE` (default). If TRUE diagnostics for cohorts in the 
cohort definition set that have not changed will be skipped and existing results 
csv files will be updated. If FALSE then diagnostics for all cohorts in the cohort 
definition set will be executed and pre-existing results files will be deleted.}

\item{incrementalFolder}{If \code{incremental = TRUE}, specify a folder where records are kept
of which cohort diagnostics has been executed. If not specified, a file named `incremental`  will be created inside the 
\code{export_folder} directory.}
}
\description{
Generates the `visit_context.csv` which contains the counts for the subjects by `cohort_id`,
`visit_concept_id` and `visit_context`. The `visit_context` categorizes visit occurrences of
subjects based on how each the start and end date of each visit related to the cohort start date
to which each subject belongs. No output will be generated for cohorts with no subjects.If there
is no cohort with subjects execution will halt and `visit_context.csv` will not be generated.
}
