% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisitContextDiagnostic.R
\name{runVisitContextDiagnostic}
\alias{runVisitContextDiagnostic}
\title{Run visit context diagnostic}
\usage{
runVisitContextDiagnostic(
  context,
  cohortDefinitionSet = NULL,
  cohortIds = NULL
)
}
\arguments{
\item{context}{A DiagnosticsContext object.}

\item{cohortDefinitionSet}{Data.frame of cohorts must include columns cohortId, cohortName, json, sql}

\item{cohortIds}{Optionally, provide a subset of cohort IDs to restrict the diagnostics to.}
}
\value{
Invisible NULL
}
\description{
Runs the visit context diagnostic for a set of cohorts.
}
