% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APACStudyPeking2025.R
\docType{package}
\name{APACStudyPeking2025-package}
\alias{APACStudyPeking2025}
\alias{APACStudyPeking2025-package}
\title{APACStudyPeking2025: A Study Package for Generating Cohort Diagnostics}
\description{
A package that generates cohort diagnostics for cohort definitions included in the package.
}
\keyword{internal}
