% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsDataModel.R
\name{subsetResultsZip}
\alias{subsetResultsZip}
\title{Subset a previously generated cohort diagnostics results zip file}
\usage{
subsetResultsZip(inputFolder, outputFolder, cohortIds)
}
\arguments{
\item{inputFolder}{Location of previously generated cohort diagnostics results zip file.}

\item{outputFolder}{Location for the output zip file.}

\item{cohortIds}{An array of one or more cohort ids}
}
\description{
Subset a previously generated cohort diagnostics results zip file sourced from a single data source by cohort ids
}
