% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TakePackageDependencySnapshot.R
\name{takepackageDependencySnapshot}
\alias{takepackageDependencySnapshot}
\title{Take a snapshot of the R environment}
\usage{
takepackageDependencySnapshot()
}
\value{
A data frame listing all the dependencies of the root package and their version numbers, in the
order in which they should be installed.
}
\description{
Take a snapshot of the R environment
}
\details{
This function records all versions used in the R environment as used by runCohortDiagnostics.
This function was borrowed from OhdsiRTools
}
