% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsDataModel.R
\name{uploadPrintFriendly}
\alias{uploadPrintFriendly}
\title{Upload print-friendly cohort representations to the database server.}
\usage{
uploadPrintFriendly(connectionDetails = NULL, schema)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package.}

\item{schema}{The schema on the postgres server where the results are stored.}
}
\description{
For all the cohorts in the 'cohort' table, this will generate print-friendly text and store it
in a new table called 'cohort_extra'.
}
