% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConceptIds.R
\name{writeOmopvocabularyTables}
\alias{writeOmopvocabularyTables}
\title{Get a copy of omop vocabulary as csv}
\usage{
writeOmopvocabularyTables(
  connectionDetails = NULL,
  connection = NULL,
  vocabularyDatabaseSchema = NULL,
  conceptIds = NULL,
  vocabularyTableNames = c("concept", "conceptAncestor", "conceptClass",
    "conceptRelationship", "conceptSynonym", "domain", "relationship", "vocabulary"),
  exportFolder
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{vocabularyDatabaseSchema}{databaseSchema where the omop vocabulary files are located.
These are most commonly the same as cdmDatabaseSchema.}

\item{conceptIds}{(optional) A vector of conceptIds to filter the extract of omop 
vocabulary files. If NULL, all conceptIds are extracted.}

\item{vocabularyTableNames}{(optional) A vector of omop vocabulary table names to download.}

\item{exportFolder}{The folder where the output is exported by Cohort Diagnostics.
If this folder does not exist, or does not have the searched file
the function will return an error.}
}
\value{
NULL. The function writes the vocabulary tables into the export folder as csv.
}
\description{
For a given list of conceptId's get a subset of omop vocabulary
of these conceptIds. These are written as csv in the export folder
}
