% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortLevelDiagnostics.R
\name{breakDownIndexEvents}
\alias{breakDownIndexEvents}
\title{Check source codes used in a cohort definition}
\usage{
breakDownIndexEvents(connectionDetails, cdmDatabaseSchema,
  oracleTempSchema = NULL, cohortDatabaseSchema = cdmDatabaseSchema,
  cohortTable = "cohort", baseUrl = NULL, cohortId = NULL,
  cohortJson = NULL, cohortSql = NULL, createCohortTable = FALSE,
  instantiateCohort = createCohortTable,
  instantiatedCohortId = cohortId)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{oracleTempSchema}{Should be used in Oracle to specify a schema where the user has write
priviliges for storing temporary tables.}

\item{baseUrl}{The base URL for the WebApi instance, for example: "http://server.org:80/WebAPI".}

\item{cohortId}{The ID of the cohort in the WebAPI instance.}

\item{cohortJson}{A characteric string containing the JSON of a cohort definition.}

\item{cohortSql}{The OHDSI SQL representation of the same cohort definition.}

\item{createCohortTable}{A logical value indicating whether a new cohort table should be created. If it
already exists it will be deleted first. If \code{createCohortTable = FALSE}, the
cohort table is assumed to already exits.}

\item{instantiateCohort}{A logical value indicating whether the cohort should be instantiated in the cohort
table. If not, the cohort is assumed to already exists.}

\item{instantiatedCohortId}{The cohort definition ID used in the cohort table.}

\item{byMonth}{Compute counts by month? If FALSE, only overall counts are computed.}
}
\value{
A data frame with concepts, and per concept the count of how often the concept was encountered at the index date.
}
\description{
This function first extracts all concept sets used in a cohort definition. Then, for each concept set
the concept found in the CDM database the contributing source codes are identified.

There are two ways to call this function:
\itemize{
  \item \code{findIncludedSourceCodes(connectionDetails, cdmDatabaseSchema, oracleTempSchema, baseUrl, cohortId)}
  \item \code{findIncludedSourceCodes(connectionDetails, cdmDatabaseSchema, oracleTempSchema, cohortJson, cohortSql)}
}
}
