% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortDiagnostics.R
\docType{package}
\name{CohortDiagnostics-package}
\alias{CohortDiagnostics}
\alias{CohortDiagnostics-package}
\title{CohortDiagnostics: Diagnostics for OHDSI Cohorts}
\description{
Diagnostics for cohorts that use the OMOP Common Data Model and the OHDSI tools.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ohdsi.github.io/CohortDiagnostics}
  \item \url{https://github.com/OHDSI/CohortDiagnostics}
  \item Report bugs at \url{https://github.com/OHDSI/CohortDiagnostics/issues}
}

}
\author{
\strong{Maintainer}: Gowtham Rao \email{rao@ohdsi.org}

Authors:
\itemize{
  \item Martijn Schuemie \email{schuemie@ohdsi.org}
  \item Patrick Ryan
  \item James Weaver
  \item Jamie Gilbert
}

Other contributors:
\itemize{
  \item Observational Health Data Science and Informatics [copyright holder]
}

}
\keyword{internal}
