% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortComparisonDiagnostics.R
\name{computeCohortOverlap}
\alias{computeCohortOverlap}
\title{Compute overlap between two cohorts}
\usage{
computeCohortOverlap(
  connectionDetails = NULL,
  connection = NULL,
  cohortDatabaseSchema,
  cohortTable = "cohort",
  targetCohortId,
  comparatorCohortId
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{cohortDatabaseSchema}{Schema name where your cohort table resides. Note that for SQL Server,
this should include both the database and schema name, for example
'scratch.dbo'.}

\item{cohortTable}{Name of the cohort table.}

\item{targetCohortId}{The cohort definition ID used to reference the target cohort in the
cohort table.}

\item{comparatorCohortId}{The cohort definition ID used to reference the comparator cohort in the
cohort table.}
}
\value{
A data frame with overlap statistics.
}
\description{
Computes the overlap between a target and a comparator cohort.
}
