% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortOverlapDiagnostics.R
\name{computeCohortOverlap}
\alias{computeCohortOverlap}
\title{Given two sets of cohorts get overlap between the cohorts.}
\usage{
computeCohortOverlap(
  connectionDetails = NULL,
  connection = NULL,
  cohortDatabaseSchema,
  cohortTable = "cohort",
  targetCohortIds,
  comparatorCohortIds,
  batchSize = 50
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{cohortDatabaseSchema}{Schema name where your cohort table resides. Note that for SQL Server,
this should include both the database and schema name, for example
'scratch.dbo'.}

\item{cohortTable}{Name of the cohort table.}

\item{targetCohortIds}{List of cohort ids that represent target cohorts}

\item{comparatorCohortIds}{List of cohort ids that represent comparator cohorts}

\item{batchSize}{{Optional, default set to 50} If running diagnostics on large set
of cohorts, this function allows you to batch them into chunks that
by default run over 50 target cohorts (and all comparator cohorts).}
}
\description{
Given two sets of cohorts, get data on overlap between the cohorts.
Note: only the first occurrence of subject_id in the cohort is used.
}
