% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConceptCountsTable.R
\name{createConceptCountsTable}
\alias{createConceptCountsTable}
\title{createConceptCountsTable}
\usage{
createConceptCountsTable(
  connectionDetails = NULL,
  connection = NULL,
  cdmDatabaseSchema,
  tempEmulationSchema = NULL,
  conceptCountsDatabaseSchema,
  conceptCountsTable = "concept_counts",
  conceptCountsTableIsTemp = FALSE
)
}
\arguments{
\item{connection}{}

\item{cdmDatabaseSchema}{}

\item{tempEmulationSchema}{}

\item{conceptCountsDatabaseSchema}{}

\item{conceptCountsTable}{}

\item{conceptCountsTableIsTemp}{}
}
\description{
Create a table containing the concept counts.
}
