% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConceptSetUtils.R
\name{createConceptCountsTable}
\alias{createConceptCountsTable}
\title{Create the concept counts table used by CohortDiagnostics}
\usage{
createConceptCountsTable(
  connection = NULL,
  cdmDatabaseSchema,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  conceptCountsDatabaseSchema = NULL,
  conceptCountsTable = "concept_counts",
  achillesDatabaseSchema = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not truly support temp tables. To emulate temp 
tables, provide a schema with write privileges where temp tables can be created.}

\item{conceptCountsDatabaseSchema}{character. The schema where the concept counts table should be created.}

\item{conceptCountsTable}{character. The name of concept counts table. If the name starts with # then a 
temporary table will be created.}

\item{achillesDatabaseSchema}{character. Schema where Achilles tables are stored in the database. 
If provided and the correct analysis IDs exist
then the Achilles results will be used to create the concept counts table.}

\item{overwrite}{logical. Should the table be overwritten if it already exists? TRUE or FALSE (default)}

\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}
}
\description{
CohortDiagnostics requires the record and person counts for all concepts in the CDM.
createConceptCountsTable will create this table. To speed up execution pre-computed 
Achilles data can be used in the creation of the table. This table can also be persisted 
across executions of CohortDiagnostics and does not need to be recreated with each run.
}
\examples{
\dontrun{
connectionDetails <- Eunomia::getEunomiaConnectionDetails()
connection <- DatabaseConnector::connect(connectionDetails)

createConceptCountsTable(connection = connection,
                         cdmDatabaseSchema = "main",
                         conceptCountsDatabaseSchema = "main",
                         conceptCountsTable = "concept_counts",
                         achillesDatabaseSchema = NULL,
                         overwrite = FALSE)  

DatabaseConnector::disconnect(connection)
}
}
