% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConceptCountsTable.R
\name{createConceptCountsTable}
\alias{createConceptCountsTable}
\title{createConceptCountsTable}
\usage{
createConceptCountsTable(
  connectionDetails = NULL,
  connection = NULL,
  cdmDatabaseSchema,
  tempEmulationSchema = NULL,
  conceptCountsTable = "concept_counts",
  conceptCountsDatabaseSchema = cdmDatabaseSchema,
  conceptCountsTableIsTemp = FALSE,
  removeCurrentTable = TRUE
)
}
\arguments{
\item{conceptCountsDatabaseSchema}{schema name for the concept counts table}

\item{conceptCountsTableIsTemp}{boolean to indicate if it should be a temporary table}

\item{removeCurrentTable}{if the current table should be removed}
}
\description{
Create a table containing concept counts. 
CohortDiagnostics performs this task in every run and takes a significant amount of time. 
However, with this function, the user can create this table beforehand and 
save it in the writing schema for further use.
}
