% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaDataDiagnostics.R
\name{createConceptCountsTable}
\alias{createConceptCountsTable}
\title{Title}
\usage{
createConceptCountsTable(connectionDetails = NULL, connection = NULL,
  cdmDatabaseSchema, workDatabaseSchema = cdmDatabaseSchema,
  conceptCountsTable = "concept_counts")
}
\arguments{
\item{connectionDetails}{}

\item{cdmDatabaseSchema}{}

\item{workDatabaseSchema}{}

\item{conceptCountsTable}{}
}
\description{
Title
}
