% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shared.R
\name{createDatabaseDataSource}
\alias{createDatabaseDataSource}
\title{Return a database data source object}
\usage{
createDatabaseDataSource(
  connection = NULL,
  connectionDetails = NULL,
  resultsDatabaseSchema,
  vocabularyDatabaseSchema = resultsDatabaseSchema
)
}
\arguments{
\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{resultsDatabaseSchema}{Schema name where the output of your Cohort Diagnostics result set it uploaded. This 
is commonly uploaded using \code{CohortDiagnostics::uploadResults} function. The
schema may be initiated using \code{CohortDiagnostics::resultsDatabaseSchema}.}

\item{vocabularyDatabaseSchema}{Schema name where your OMOP vocabulary data resides. This is 
commonly the same as cdmDatabaseSchema. Note that for 
SQL Server, this should include both the database and
schema name, for example 'vocabulary.dbo'.}
}
\value{
Returns a list with information on database data source
}
\description{
Collects a list of objects needed to connect to a database datsource. This includes one of
\code{DatabaseConnector::createConnectionDetails} object, or a DBI database connection created
using either \code{DatabaseConnector::connection} or \code{pool::dbPool}, and a names of
resultsDatabaseSchema and vocabularyDatabaseSchema
}
