% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsDataModelQueries.R
\name{createFileDataSource}
\alias{createFileDataSource}
\title{Return a file data source object}
\usage{
createFileDataSource(premergedDataFile, envir = .GlobalEnv)
}
\arguments{
\item{premergedDataFile}{an .RData/rds object the output
of \code{CohortDiagnostics::preMergeDiagnosticsFiles}}

\item{envir}{(optional) R-environment to read premerged data. By default this is the
global environment.}
}
\value{
R environment containing data conforming to Cohort Diagnostics results data model specifications.
}
\description{
Given a premerged file (an .RData/rds object the output
of \code{CohortDiagnostics::preMergeDiagnosticsFiles} reads the object into
memory and makes it available for query.
}
