% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsDataModel.R
\name{createResultsDataModel}
\alias{createResultsDataModel}
\title{Create the results data model tables on a database server.}
\usage{
createResultsDataModel(connection = NULL, connectionDetails = NULL, schema)
}
\arguments{
\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{schema}{The schema on the postgres server where the tables will be created.}
}
\description{
Create the results data model tables on a database server.
}
\details{
Only PostgreSQL servers are supported.
}
