% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunDiagnostics.R
\name{executeDiagnosticsCdm}
\alias{executeDiagnosticsCdm}
\title{executeDiagnosticsCdm}
\usage{
executeDiagnosticsCdm(
  cdm,
  cohortDefinitionSet,
  cohortTable = "cohort",
  exportFolder,
  minCellCount = 5,
  runInclusionStatistics = TRUE,
  runIncludedSourceConcepts = TRUE,
  runOrphanConcepts = TRUE,
  runTimeSeries = TRUE,
  runVisitContext = TRUE,
  runBreakdownIndexEvents = TRUE,
  runIncidenceRate = TRUE,
  runCohortRelationship = TRUE,
  runTemporalCohortCharacterization = TRUE
)
}
\arguments{
\item{cdm}{}

\item{cohortDefinitionSet}{}

\item{cohortTable}{}

\item{exportFolder}{}

\item{minCellCount}{}

\item{runInclusionStatistics}{}

\item{runIncludedSourceConcepts}{}

\item{runOrphanConcepts}{}

\item{runTimeSeries}{}

\item{runVisitContext}{}

\item{runBreakdownIndexEvents}{}

\item{runIncidenceRate}{}

\item{runCohortRelationship}{}

\item{runTemporalCohortCharacterization}{}
}
\description{
executeDiagnosticsCdm
}
