% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseFunctions.R
\name{executeSql}
\alias{executeSql}
\title{executeSql}
\usage{
executeSql(
  connection,
  sql,
  profile = FALSE,
  progressBar = !as.logical(Sys.getenv("TESTTHAT", unset = FALSE)),
  reportOverallTime = TRUE,
  errorReportFile = file.path(getwd(), "errorReportSql.txt"),
  runAsBatch = FALSE
)
}
\arguments{
\item{connection}{}

\item{sql}{}

\item{profile}{}

\item{progressBar}{}

\item{reportOverallTime}{}

\item{errorReportFile}{}

\item{runAsBatch}{}
}
\description{
executeSql
}
