% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shared.R
\name{getCirceRenderedExpression}
\alias{getCirceRenderedExpression}
\title{Returns list with circe generated documentation}
\usage{
getCirceRenderedExpression(
  cohortDefinition,
  cohortName = NULL,
  embedText = NULL,
  includeConceptSets = FALSE
)
}
\arguments{
\item{cohortDefinition}{An R object (list) with a list representation of the cohort definition expression.}

\item{cohortName}{Name for the cohort definition}

\item{embedText}{Any additional text to embed in the top of circe generated content.}

\item{includeConceptSets}{Do you want to inclued concept set in the documentation}
}
\value{
list object
}
\description{
Returns list with circe generated documentation
}
