% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shared.R
\name{getCirceRenderedExpression}
\alias{getCirceRenderedExpression}
\title{Returns list with circe generated documentation}
\usage{
getCirceRenderedExpression(cohortDefinition)
}
\arguments{
\item{cohortDefinitionExpression}{An object with a list representation of the cohort definition expression.}
}
\value{
list object
}
\description{
Returns list with circe generated documentation
}
