% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortCharacterizationDiagnostics.R
\name{getCohortCharacteristics}
\alias{getCohortCharacteristics}
\title{Get Characteristics for a cohort}
\usage{
getCohortCharacteristics(
  connectionDetails = NULL,
  connection = NULL,
  cdmDatabaseSchema,
  tempEmulationSchema = NULL,
  cohortDatabaseSchema = cdmDatabaseSchema,
  cohortTable = "cohort",
  cohortIds = NULL,
  cdmVersion = 5,
  covariateSettings = createDefaultCovariateSettings(),
  batchSize = 100
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not truly support temp tables. To emulate temp 
tables, provide a schema with write privileges where temp tables can be created.}

\item{cohortDatabaseSchema}{Schema name where your cohort table resides. Note that for SQL Server,
this should include both the database and schema name, for example
'scratch.dbo'.}

\item{cohortTable}{Name of the cohort table.}

\item{cohortIds}{Optionally, provide a subset of cohort IDs to restrict the
diagnostics to.}

\item{cdmVersion}{The version of the OMOP CDM. Default 5. (Note: only 5 is supported.)}

\item{covariateSettings}{Either an object of type \code{covariateSettings} as created using one of
the createCovariateSettings (createTemporalCovariateSettings if temporal
characterization) function in the FeatureExtraction package, or a list
of such objects. If unspecified, default covariate settings as specified
by FeatureExtraction is computed, this is sufficient for presenting default
table 1. See documentation of FeatureExtraction on how to specify
CovariateSettings object.}

\item{batchSize}{{Optional, default set to 100} If running characterization on larget set
of cohorts, this function allows you to batch them into chunks that run
as a batch.}

\item{exportFolder}{The folder where the output will be exported to. If this folder
does not exist it will be created.}
}
\description{
Given a set of instantiated cohorts get Characteristics for the cohort using \code{FeatureExtraction::getDbCovariateData}.

If runTemporalCohortCharacterization argument is TRUE, then the following default covariateSettings object will be created
using \code{RFeatureExtraction::createTemporalCovariateSettings}.
}
