% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortCharacterizationDiagnostics.R
\name{getCohortCharacteristics}
\alias{getCohortCharacteristics}
\title{Get characterization for cohorts}
\usage{
getCohortCharacteristics(
  connectionDetails = NULL,
  connection = NULL,
  cdmDatabaseSchema,
  tempEmulationSchema = NULL,
  cohortDatabaseSchema = cdmDatabaseSchema,
  cohortTable = "cohort",
  cohortIds,
  cdmVersion = 5,
  temporalCovariateSettings,
  exportFolder
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not truly support temp tables. To emulate temp 
tables, provide a schema with write privileges where temp tables can be created.}

\item{cohortDatabaseSchema}{Schema name where your user has write access (has CRUD privileges). This is the location,
of the cohort tables. This is also the location for the optional table such as 
the unique_concept_id table. 
In incremental mode: It is assumed that the tables in this location have not been modified 
outside this application, and the content was written by this application for current 
project only. A good practice would be to ensure that the database table names are unique 
for the project. e.g. 'EPI101_COHORT', 'EPI101_UNIQUE_CONCEPT_ID' (please try to keep 
length < 22). Another assumption: the content of these tables were written by this application
(i.e. they were not altered outside of this application) and the tables are available during the
life cycle of this project. The local files that are created by the application during incremental
mode such as the 'recordKeepingFile'.
Note that for SQL Server, this should include both the database and schema name, for example
'scratch.dbo'.}

\item{cohortTable}{Name of the cohort table.}

\item{cohortIds}{cohort ids to identify the cohort definitions to use on FeatureExtraction}

\item{cdmVersion}{The version of the OMOP CDM. Default 5. (Note: only 5 is supported.)}

\item{temporalCovariateSettings}{Either an object of type \code{covariateSettings} as created using one of
the createTemporalCovariateSettings function in the FeatureExtraction package, or a list
of such objects.}

\item{exportFolder}{The folder where the output will be exported to. If this folder
does not exist it will be created.}
}
\description{
Using FeatureExtraction, this function gets temporal characteristics for the cohorts
under assessment in cohort diagnostics.
}
