% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shared.R
\name{getConceptMetadata}
\alias{getConceptMetadata}
\title{Returns a metadata for concept ids}
\usage{
getConceptMetadata(
  dataSource,
  databaseIds = NULL,
  cohortIds = NULL,
  vocabularyDatabaseSchema = NULL,
  conceptIds = NULL,
  getDatabaseMetadata = TRUE,
  getConceptRelationship = TRUE,
  getConceptAncestor = TRUE,
  getConceptSynonym = TRUE,
  getConceptCount = TRUE,
  getConceptCooccurrence = TRUE,
  getIndexEventCount = TRUE,
  getConceptMappingCount = TRUE,
  getFixedTimeSeries = TRUE,
  getRelativeTimeSeries = TRUE
)
}
\arguments{
\item{dataSource}{A list object that is the output of \code{createDatabaseDataSource} or \code{createFileDataSource} function. 
This object helps direct the function to query data from the database (created by \code{createDatabaseDataSource})
or a local premerged file (created by \code{createFileDataSource}). Premerged files are output of cohortDiagnostics
compiled into RData using \code{preMergeDiagnosticsFiles}. Database DataSources are data inserted into a remote
database (only a postgres database is supported) with tables created with DDL function \code{createResultsDataModel}
and uploaded using \code{uploadResults}}

\item{databaseIds}{A vector one or more databaseIds to retrieve the results for. This is a 
character field values from the `databaseId` field of the `database` table 
of the results data model.}

\item{cohortIds}{A vector of one or more Cohort Ids.}

\item{vocabularyDatabaseSchema}{Schema name where your OMOP vocabulary data resides. This is 
commonly the same as cdmDatabaseSchema. Note that for 
SQL Server, this should include both the database and
schema name, for example 'vocabulary.dbo'.}

\item{conceptIds}{A vector of one or more Concept Ids.}

\item{getDatabaseMetadata}{Do you want to get database metadata i.e. person/records/person days in observation table}

\item{getConceptRelationship}{Do you want conceptRelationship?}

\item{getConceptAncestor}{Do you want conceptAncestor?}

\item{getConceptSynonym}{Do you want conceptSynonym?}

\item{getConceptCount}{Do you want conceptCount?}

\item{getConceptCooccurrence}{Do you want concept cooccurrence?}

\item{getIndexEventCount}{Do you want index event concept count?}

\item{getConceptMappingCount}{Do you want concept mapping count?}

\item{getFixedTimeSeries}{Do you want a conceptIds database level time series data reported on actual dates?}

\item{getRelativeTimeSeries}{Do you want cohort level time series data reported relative to cohort start date?}
}
\value{
Returns a list of data frames (tibbles)
}
\description{
Returns a metadata for a given list of concept ids that includes concept synonyms,
concept relationship, concept ancestor, concept count per database,
concept cooccurrence on index date per database and cohortId.
}
