% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executeDiagnostics.R
\name{getDefaultCovariateSettings}
\alias{getDefaultCovariateSettings}
\title{Get default covariate settings}
\usage{
getDefaultCovariateSettings()
}
\value{
Default covariate settings
}
\description{
Default covariate settings for cohort diagnostics execution
}
