% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shared.R
\name{getDomainInformation}
\alias{getDomainInformation}
\title{Get domain information}
\usage{
getDomainInformation(versionNumber = NULL, packageName = NULL)
}
\arguments{
\item{versionNumber}{Which version of Cohort Diagnostics. Default will be the most recent version.}

\item{packageName}{e.g. 'CohortDiagnostics'}
}
\value{
A tibble data frame object with domain information
}
\description{
Get domain information
}
