% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IncidenceRates.R, R/ResultsPulls.R
\name{getIncidenceRate}
\alias{getIncidenceRate}
\title{Compute incidence rate for a cohort}
\usage{
getIncidenceRate(
  connection = NULL,
  connectionDetails = NULL,
  cohortId,
  databaseIds,
  stratification,
  minPersonYears = 1000,
  resultsDatabaseSchema = NULL
)

getIncidenceRate(
  connection = NULL,
  connectionDetails = NULL,
  cohortId,
  databaseIds,
  stratification,
  minPersonYears = 1000,
  resultsDatabaseSchema = NULL
)
}
\arguments{
\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{cohortId}{Cohort Id to retrieve the data. This is one of the integer (bigint) value from
cohortId field in cohort table of the results data model.}

\item{databaseIds}{A vector one or more databaseIds to retrieve the results for. This is a character 
field value from the databaseId field in the database table of the results data model.}

\item{minPersonYears}{(optional) Default value = 1000. Minimum person years needed to create plot.}

\item{resultsDatabaseSchema}{(optional) The databaseSchema where the results data model of cohort diagnostics
is stored. This is only required when \code{connectionDetails} or 
\code{\link[DatabaseConnector]{connect}} is provided.}

\item{cohortDatabaseSchema}{Schema name where your cohort table resides. Note that for SQL Server,
this should include both the database and schema name, for example
'scratch.dbo'.}

\item{cohortTable}{Name of the cohort table.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{cdmVersion}{The version of the OMOP CDM. Default 5. (Note: only 5 is supported.)}

\item{oracleTempSchema}{Should be used in Oracle to specify a schema where the user has write
privileges for storing temporary tables.}

\item{firstOccurrenceOnly}{Use only the first occurrence of the cohort per person?}

\item{washoutPeriod}{The minimum amount of observation time required before the occurrence
of a cohort entry. This is also used to eliminate immortal time from
the denominator.}
}
\value{
Returns a data frame of cohort count, person year count, and incidence rate per 1000 persons
years with the following stratifications: 1) no stratification, 2) gender stratification,
3) age (10-year) stratification, 4) calendar year and age (10-year) stratification, 5) calendar
year and gender stratification, 6) calendar year, age (10-year), and gender stratification with
option to save dataframes.

The function will return a tibble data frame object.
}
\description{
Returns yearly incidence rate stratified by age and gender

Get incidence rate data for plotting from data stored in cohort diagnostics results data
model. The output of this function may be used to create plots or tables
related to incidence rate diagnostics. This function will look for incidence_rate table
in results data model.

Note: this function relies on data available in Cohort Diagnostics results data model. There are
two methods to connect to the results data model, database mode and in-memory mode.

Database mode: If either \code{connectionDetails} or \code{\link[DatabaseConnector]{connect}} is 
provided in the function call, the query is set to database mode. In the absence of both 
\code{connectionDetails} and \code{\link[DatabaseConnector]{connect}}, the query will be in-memory 
mode. In in-memory mode, R will expect the data in results data model to be available in R's memory.
In database mode, R will perform a database query. Objects in R's memory are expected to
follow camelCase naming conventions, while objects in dbms are expected to follow snake-case naming
conventions. In database mode, vocabulary tables (if needed) are used from vocabularySchema (which
defaults to resultsDatabaseSchema.). In in-memory mode, vocabulary tables are assumed to in R's 
memory.
}
\examples{
\dontrun{
incidenceRate <- getIncidenceRate(cohortId = 343242,
                                  databaseIds = c('eunomia', 'hcup'))
}

}
