% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shared.R
\name{getResultsDataModelSpecifications}
\alias{getResultsDataModelSpecifications}
\title{Get specifications for Cohort Diagnostics results data model}
\usage{
getResultsDataModelSpecifications(versionNumber = NULL, packageName = NULL)
}
\arguments{
\item{versionNumber}{Which version of Cohort Diagnostics. Default will be the most recent version.}

\item{packageName}{e.g. 'CohortDiagnostics'}
}
\value{
A tibble data frame object with specifications
}
\description{
Get specifications for Cohort Diagnostics results data model
}
