% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsDataModelQueries.R
\name{getResultsFromTimeSeries}
\alias{getResultsFromTimeSeries}
\title{Returns data from time_series table of Cohort Diagnostics results data model}
\usage{
getResultsFromTimeSeries(dataSource, cohortIds = NULL, databaseIds = NULL)
}
\arguments{
\item{dataSource}{A list object that is the output of \code{createDatabaseDataSource} or \code{createFileDataSource} function. 
This object helps direct the function to query data from the database (created by \code{createDatabaseDataSource})
or a local premerged file (created by \code{createFileDataSource}). Premerged files are output of cohortDiagnostics
compiled into RData using \code{preMergeDiagnosticsFiles}. Database DataSources are data inserted into a remote
database (only a postgres database is supported) with tables created with DDL function \code{createResultsDataModel}
and uploaded using \code{uploadResults}}

\item{cohortIds}{A vector of one or more Cohort Ids.}

\item{databaseIds}{A vector one or more databaseIds to retrieve the results for. This is a 
character field values from the `databaseId` field of the `database` table 
of the results data model.}
}
\value{
Returns a list of tsibble (time series) objects with results that conform to time series
table in Cohort Diagnostics results data model. There are three list objects, labelled
m for monthly, q for quarterly and y for yearly. The periodBegin variable is in the 
format of tsibble::yearmonth for monthly, tsibble::yearquarter for quarter and integer
for year.
}
\description{
Returns data from time_series table of Cohort Diagnostics results data model.
The returned object is a tsibble, but to use in time series analysis, gaps
need to be filled. Only absolute values are returned i.e. negative values are
converted to positives.
}
