% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseFunctions.R
\name{getTableNames}
\alias{getTableNames}
\title{getTableNames}
\usage{
getTableNames(connection, cdmDatabaseSchema)
}
\arguments{
\item{connection}{}

\item{cdmDatabaseSchema}{}
}
\description{
getTableNames
}
