% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsPulls.R
\name{getTimeDistribution}
\alias{getTimeDistribution}
\title{Get time distribution data for plotting}
\usage{
getTimeDistribution(
  connection = NULL,
  connect = NULL,
  cohortId,
  databaseId,
  resultsDatabaseSchema = NULL
)
}
\arguments{
\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{cohortId}{Cohort Id to retrieve the data. This is one of the integer (bigint) value from
cohortId field in cohort table of the results data model.}

\item{databaseId}{The database to retrieve the results for. This is the character field value from
the databaseId field in the database table of the results data model.}

\item{resultsDatabaseSchema}{(optional) The databaseSchema where the results data model of cohort diagnostics
is stored. This is only required when \code{connectionDetails} or 
\code{\link[DatabaseConnector]{connect}} is provided.}

\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}
}
\value{
The function will return a tibble data frame object.
}
\description{
Get time distribution data for plotting from data stored in cohort diagnostics results data
model. The output of this function may be used to create plots or tables
related to time distribution diagnostics. This function will look for time_distribution table
in results data model.

Note: this function relies on data available in Cohort Diagnostics results data model. There are
two methods to connect to the resutls data model, database mode and in-memory mode.

Database mode: If either \code{connectionDetails} or \code{\link[DatabaseConnector]{connect}} is 
provided in the function call, the query is set to database mode. In the absence of both 
\code{connectionDetails} and \code{\link[DatabaseConnector]{connect}}, the query will be in-memory 
mode. In in-memory mode, R will expect the data in results data model to be available in R's memory.
In database mode, R will perform a database query. Objects in R's memory are expected to
follow camelCase naming conventions, while objects in dbms are expected to follow snake-case naming
conventions. In database mode, vocabulary tables (if needed) are used from vocabularySchema (which
defaults to resultsDatabaseSchema.). In in-memory mode, vocabulary tables are assumed to in R's 
memory.
}
\examples{
\dontrun{
timeDistribution <- getTimeDistribution(cohortId = 343242,
                                        databaseId = 'eunomia')
}

}
