% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shiny.R
\name{launchCohortExplorer}
\alias{launchCohortExplorer}
\title{Launch the CohortExplorer Shiny app}
\usage{
launchCohortExplorer(
  connectionDetails,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTable,
  cohortId,
  sampleSize = 100,
  subjectIds = NULL
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{cohortDatabaseSchema}{Schema name where your cohort table resides. Note that for SQL Server,
this should include both the database and schema name, for example
'scratch.dbo'.}

\item{cohortTable}{Name of the cohort table.}

\item{cohortId}{The ID of the cohort.}

\item{sampleSize}{Number of subjects to sample from the cohort. Ignored if subjectIds is specified.}

\item{subjectIds}{A vector of subject IDs to view.}
}
\description{
Launch the CohortExplorer Shiny app
}
\details{
Launches a Shiny app that allows the user to explore a cohort of interest.
}
