% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shiny.R
\name{launchDiagnosticsExplorer}
\alias{launchDiagnosticsExplorer}
\title{Launch the Diagnostics Explorer Shiny app}
\usage{
launchDiagnosticsExplorer(
  dataFolder = "data",
  connectionDetails = NULL,
  resultsDatabaseSchema = NULL,
  vocabularyDatabaseSchema = resultsDatabaseSchema,
  cohortBaseUrl = "https://atlas.ohdsi.org/#/cohortdefinition/",
  conceptBaseUrl = "https://athena.ohdsi.org/search-terms/terms/",
  runOverNetwork = FALSE,
  port = 80,
  launch.browser = FALSE
)
}
\arguments{
\item{dataFolder}{A folder where the premerged file is stored. Use
the \code{\link{preMergeDiagnosticsFiles}} function to generate this file.}

\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package, specifying how to connect to the server where
the CohortDiagnostics results have been uploaded using the 
\code{\link{uploadResults}} function.}

\item{resultsDatabaseSchema}{The schema on the database server where the CohortDiagnostics results 
have been uploaded.}

\item{vocabularyDatabaseSchema}{The schema on the database server where the vocbulary tables are located.}

\item{cohortBaseUrl}{The base URL for constructing linkouts to an ATLAS instance, using the 
webApiCohortId in the cohortsToCreate file. If NULL, no linkouts will be 
created.}

\item{conceptBaseUrl}{The base URL for constructing linkouts to an Athena instance, using the
concept ID.}

\item{runOverNetwork}{(optional) Do you want the app to run over your network?}

\item{port}{(optional) Only used if \code{runOverNetwork} = TRUE.}

\item{launch.browser}{Should the app be launched in your default browser, or in a Shiny window.
Note: copying to clipboard will not work in a Shiny window.}
}
\description{
Launch the Diagnostics Explorer Shiny app
}
\details{
Launches a Shiny app that allows the user to explore the diagnostics
}
