% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shiny.R
\name{launchDiagnosticsExplorer}
\alias{launchDiagnosticsExplorer}
\title{Launch the Diagnostics Explorer Shiny app}
\usage{
launchDiagnosticsExplorer(
  sqliteDbPath = "MergedCohortDiagnosticsData.sqlite",
  connectionDetails = NULL,
  resultsDatabaseSchema = NULL,
  vocabularyDatabaseSchema = NULL,
  vocabularyDatabaseSchemas = resultsDatabaseSchema,
  tablePrefix = "",
  cohortTableName = "cohort",
  databaseTableName = "database",
  aboutText = NULL,
  runOverNetwork = FALSE,
  port = 80,
  launch.browser = FALSE,
  enableAnnotation = TRUE
)
}
\arguments{
\item{sqliteDbPath}{Path to merged sqlite file. See \code{\link{createMergedResultsFile}} to create file.}

\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package, specifying how to connect to the server where
the CohortDiagnostics results have been uploaded using the
\code{\link{uploadResults}} function.}

\item{resultsDatabaseSchema}{The schema on the database server where the CohortDiagnostics results
have been uploaded.}

\item{vocabularyDatabaseSchema}{(Deprecated) Please use vocabularyDatabaseSchemas.}

\item{vocabularyDatabaseSchemas}{(optional) A list of one or more schemas on the database server where the vocabulary tables are located.
The default value is the value of the resultsDatabaseSchema. We can provide a list of vocabulary schema
that might represent different versions of the OMOP vocabulary tables. It allows us to compare the impact
of vocabulary changes on Diagnostics. Not supported with an sqlite database.}

\item{tablePrefix}{(Optional)  string to insert before table names (e.g. "cd_") for database table names}

\item{cohortTableName}{(Optional) if cohort table name differs from the standard - cohort (ignores prefix if set)}

\item{databaseTableName}{(Optional) if database table name differs from the standard - database (ignores prefix if set)}

\item{aboutText}{Text (using HTML markup) that will be displayed in an About tab in the Shiny app.
If not provided, no About tab will be shown.}

\item{runOverNetwork}{(optional) Do you want the app to run over your network?}

\item{port}{(optional) Only used if \code{runOverNetwork} = TRUE.}

\item{launch.browser}{Should the app be launched in your default browser, or in a Shiny window.
Note: copying to clipboard will not work in a Shiny window.}

\item{enableAnnotation}{Enable annotation functionality in shiny app}
}
\description{
Launch the Diagnostics Explorer Shiny app
}
\details{
Launches a Shiny app that allows the user to explore the diagnostics
}
