% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shiny.R
\name{launchDiagnosticsExplorer}
\alias{launchDiagnosticsExplorer}
\title{Launch the Diagnostics Explorer Shiny app}
\usage{
launchDiagnosticsExplorer(
  dataFolder = "data",
  dataFile = "PreMerged.RData",
  connectionDetails = NULL,
  resultsDatabaseSchema = NULL,
  vocabularyDatabaseSchema = resultsDatabaseSchema,
  aboutText = NULL,
  runOverNetwork = FALSE,
  port = 80,
  launch.browser = FALSE
)
}
\arguments{
\item{dataFolder}{A folder where the premerged file is stored. Use
the \code{\link{preMergeDiagnosticsFiles}} function to generate this file.}

\item{dataFile}{(Optional) The name of the .RData file with results. It is commonly known as the
Premerged file.}

\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package, specifying how to connect to the server where
the CohortDiagnostics results have been uploaded using the 
\code{\link{uploadResults}} function.}

\item{resultsDatabaseSchema}{The schema on the database server where the CohortDiagnostics results 
have been uploaded.}

\item{vocabularyDatabaseSchema}{The schema on the database server where the vocabulary tables are located.}

\item{aboutText}{Text (using HTML markup) that will be displayed in an About tab in the Shiny app.
If not provided, no About tab will be shown.}

\item{runOverNetwork}{(optional) Do you want the app to run over your network?}

\item{port}{(optional) Only used if \code{runOverNetwork} = TRUE.}

\item{launch.browser}{Should the app be launched in your default browser, or in a Shiny window.
Note: copying to clipboard will not work in a Shiny window.}
}
\description{
Launch the Diagnostics Explorer Shiny app
}
\details{
Launches a Shiny app that allows the user to explore the diagnostics
}
