% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IncidenceProportion.R
\name{plotIncidenceProportion}
\alias{plotIncidenceProportion}
\title{Generate stability plots}
\usage{
plotIncidenceProportion(incidenceProportion, panel = "age",
  restrictToFullAgeData = FALSE, fileName = NULL)
}
\arguments{
\item{incidenceProportion}{Incidence proportion time series data for plotting generated using
\code{\link{getIncidenceProportion}} function.}

\item{panel}{Create trellis panels by gender or by age? select "age" or "gender", defaults to "age"}

\item{fileName}{Optional: directory and filename of plot to be saved}
}
\value{
A list of 4 gg objects generated by the ggplot2 package with option to save as 1 *.rds
file and 4 png files.
}
\description{
Characterizes the incidence proportion of a phenotype as a time series visualization
}
\details{
Generates time series plots of the incidence proportion per 1000 persons of phenotype
entry by year, by year and 10-year age group, by year and gender, and by year and
10-year age group and gender
}
