% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IncidenceRates.R
\name{plotincidenceRate}
\alias{plotincidenceRate}
\title{Plot incidence rate by year, age, and/or gender}
\usage{
plotincidenceRate(
  incidenceRate,
  minPersonYears = 1000,
  stratifyByAge = TRUE,
  stratifyByGender = TRUE,
  stratifyByCalendarYear = TRUE,
  fileName = NULL
)
}
\arguments{
\item{minPersonYears}{Estimates get very unstable with low background counts, so removing them
makes for cleaner plots.}

\item{stratifyByAge}{Should the plot be stratified by age?}

\item{stratifyByGender}{Should the plot be stratified by gender?}

\item{stratifyByCalendarYear}{Should the plot be stratified by calendar year?}

\item{fileName}{Optional: name of the file where the plot should be saved, for
example 'plot.png'. See the function \code{ggsave} in the ggplot2
package for supported file formats.}

\item{data}{Incidence rate time series data for plotting generated using
\code{\link{getIncidenceRate}} function.}
}
\value{
A ggplot object. Use the \code{\link[ggplot2]{ggsave}} function to save to file in a different
format.
}
\description{
Characterizes the incidence rate of a cohort definition.
}
\details{
Generates time series plots of the incidence rate per 1000 person years of cohort entry by
year, age, and/or gender.
}
