% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsPlot.R
\name{plotIncidenceRate}
\alias{plotIncidenceRate}
\title{Get ggplot object with incidence rate plot.}
\usage{
plotIncidenceRate(
  data,
  cohortIds = NULL,
  databaseIds = NULL,
  stratifyByAgeGroup = TRUE,
  stratifyByGender = TRUE,
  stratifyByCalendarYear = TRUE,
  yscaleFixed = FALSE
)
}
\arguments{
\item{data}{A tibble data frame object that is the output 
of \code{\link{getIncidenceRate}} function.}

\item{cohortIds}{A vector of one or more integer (bigint) to plot.}

\item{databaseIds}{A vector of one or more databaseIds to plot.}

\item{stratifyByAgeGroup}{Do you want to stratify by age?}

\item{stratifyByGender}{Do you want to stratify by gender?}

\item{stratifyByCalendarYear}{Do you want to stratify by calendar year?}

\item{yscaleFixed}{Do you want to rescale y-axis?}
}
\value{
A ggplot object.
}
\description{
Get ggplot object with incidence rate plot.
}
\examples{
\dontrun{
incidenceRatePlot <- plotIncidenceRate(data = data)
}

}
