% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shiny.R
\name{preMergeDiagnosticsFiles}
\alias{preMergeDiagnosticsFiles}
\title{Premerge Shiny diagnostics files}
\usage{
preMergeDiagnosticsFiles(
  dataFolder,
  outputFolderForCsvAndZip = NULL,
  minCovariateProportion = 0
)
}
\arguments{
\item{dataFolder}{folder where the exported zip files for the diagnostics are stored. Use
the \code{\link{runCohortDiagnostics}} function to generate these zip files. 
Zip files containing results from multiple databases may be placed in the same
folder.}

\item{outputFolderForCsvAndZip}{(optional) folder where the post processed files for the diagnostics are to be stored. 
These files may be used with the results viewer or may be uploaded into RDMS.
Note this has to be different from dataFolder. If not provided, then only
premerged file is generated in the dataFolder. If provided, and different
from dataFolder the output will include csv, zip.}

\item{minCovariateProportion}{minimum value threshold for covariates to be included 
in premerged file (valid number (maybe decimal) between 0 to 1)}
}
\description{
If there are many diagnostics files, starting the Shiny app may take a very long time. This function 
already does most of the preprocessing, increasing loading speed.

The merged data will be stored in the same folder, and will automatically be recognized by the Shiny app.
}
