% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shiny.R
\name{preMergeDiagnosticsFiles}
\alias{preMergeDiagnosticsFiles}
\title{Premerge Shiny diagnostics files}
\usage{
preMergeDiagnosticsFiles(dataFolder, tempFolder = tempdir())
}
\arguments{
\item{dataFolder}{folder where the exported zip files for the diagnostics are stored. Use
the \code{\link{runCohortDiagnostics}} function to generate these zip files.
Zip files containing results from multiple databases may be placed in the same
folder.}

\item{tempFolder}{A folder on the local file system where the zip files are extracted to. Will be cleaned
up when the function is finished. Can be used to specify a temp folder on a drive that
has sufficient space if the default system temp space is too limited.}
}
\description{
This function combines diagnostics results from one or more databases into a single file. The result is a
single file that can be used as input for the Diagnostics Explorer Shiny app.

It also checks whether the results conform to the results data model specifications.
}
