% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseFunctions.R
\name{querySql}
\alias{querySql}
\title{querySql}
\usage{
querySql(
  connection,
  sql,
  errorReportFile = file.path(getwd(), "errorReportSql.txt"),
  snakeCaseToCamelCase = FALSE,
  integerAsNumeric = getOption("databaseConnectorIntegerAsNumeric", default = TRUE),
  integer64AsNumeric = getOption("databaseConnectorInteger64AsNumeric", default = TRUE)
)
}
\arguments{
\item{connection}{}

\item{sql}{}

\item{errorReportFile}{}

\item{snakeCaseToCamelCase}{}

\item{integerAsNumeric}{}

\item{integer64AsNumeric}{}
}
\description{
querySql
}
