% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseFunctions.R
\name{renderTranslateQuerySql}
\alias{renderTranslateQuerySql}
\title{renderTranslateQuerySql}
\usage{
renderTranslateQuerySql(
  connection,
  sql,
  errorReportFile = file.path(getwd(), "errorReportSql.txt"),
  snakeCaseToCamelCase = FALSE,
  oracleTempSchema = NULL,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  integerAsNumeric = getOption("databaseConnectorIntegerAsNumeric", default = TRUE),
  integer64AsNumeric = getOption("databaseConnectorInteger64AsNumeric", default = TRUE),
  ...
)
}
\arguments{
\item{connection}{db connection}

\item{sql}{The SQL to be executed}

\item{errorReportFile}{error file}

\item{snakeCaseToCamelCase}{snake case to camel case boolean}

\item{oracleTempSchema}{oracle temp schema}

\item{tempEmulationSchema}{temp emulation schema}

\item{integerAsNumeric}{int as numeric boolean}

\item{integer64AsNumeric}{int64 as numeric boolean}

\item{...}{parameters that will be used to render the SQL.}
}
\value{
the query results
}
\description{
renderTranslateQuerySql
}
