% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBreakdownIndexEvents.R
\name{runBreakdownIndexEvents}
\alias{runBreakdownIndexEvents}
\title{Title}
\usage{
runBreakdownIndexEvents(
  connection,
  tempEmulationSchema,
  cdmDatabaseSchema,
  vocabularyDatabaseSchema = cdmDatabaseSchema,
  databaseId,
  cohorts,
  runIncludedSourceConcepts,
  runOrphanConcepts,
  runBreakdownIndexEvents,
  exportFolder,
  minCellCount,
  conceptCountsDatabaseSchema = NULL,
  conceptCountsTable = "concept_counts",
  conceptCountsTableIsTemp = FALSE,
  cohortDatabaseSchema,
  cohortTable,
  useExternalConceptCountsTable = FALSE,
  incremental = FALSE,
  conceptIdTable = NULL,
  recordKeepingFile,
  useAchilles,
  resultsDatabaseSchema
)
}
\arguments{
\item{connection}{}

\item{tempEmulationSchema}{}

\item{cdmDatabaseSchema}{}

\item{vocabularyDatabaseSchema}{}

\item{databaseId}{}

\item{cohorts}{}

\item{runIncludedSourceConcepts}{}

\item{runOrphanConcepts}{}

\item{runBreakdownIndexEvents}{}

\item{exportFolder}{}

\item{minCellCount}{}

\item{conceptCountsDatabaseSchema}{}

\item{conceptCountsTable}{}

\item{conceptCountsTableIsTemp}{}

\item{cohortDatabaseSchema}{}

\item{cohortTable}{}

\item{useExternalConceptCountsTable}{}

\item{incremental}{}

\item{conceptIdTable}{}

\item{recordKeepingFile}{}

\item{useAchilles}{}

\item{resultsDatabaseSchema}{}
}
\description{
Title
}
