% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCohortRelationship.R
\name{runCohortRelationship}
\alias{runCohortRelationship}
\title{Title}
\usage{
runCohortRelationship(
  connection,
  databaseId,
  exportFolder,
  cohortDatabaseSchema,
  cdmDatabaseSchema,
  tempEmulationSchema,
  cohortTable,
  cohortDefinitionSet,
  temporalCovariateSettings,
  minCellCount,
  recordKeepingFile,
  incremental,
  batchSize = getOption("CohortDiagnostics-Relationship-batch-size", default = 500)
)
}
\arguments{
\item{connection}{}

\item{databaseId}{}

\item{exportFolder}{}

\item{cohortDatabaseSchema}{}

\item{cdmDatabaseSchema}{}

\item{tempEmulationSchema}{}

\item{cohortTable}{}

\item{cohortDefinitionSet}{}

\item{temporalCovariateSettings}{}

\item{minCellCount}{}

\item{recordKeepingFile}{}

\item{incremental}{}

\item{batchSize}{}
}
\description{
Title
}
