% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConceptSetDiagnostics.R
\name{runConceptSetDiagnostics}
\alias{runConceptSetDiagnostics}
\title{Run concept set diagnostics}
\usage{
runConceptSetDiagnostics(
  connection = NULL,
  connectionDetails = NULL,
  tempEmulationSchema = NULL,
  cdmDatabaseSchema,
  vocabularyDatabaseSchema = cdmDatabaseSchema,
  cohorts,
  cohortIds = NULL,
  cohortDatabaseSchema = NULL,
  cohortTable = NULL,
  keep2BillionConceptId = FALSE,
  runConceptSetOptimization = TRUE,
  runExcludedConceptSet = TRUE,
  runOrphanConcepts = TRUE,
  runBreakdownIndexEvents = TRUE,
  runBreakdownIndexEventRelativeDays = c(0),
  runIndexDateConceptCoOccurrence = FALSE,
  runStandardToSourceMappingCount = TRUE,
  runConceptCount = TRUE,
  runConceptCountByCalendarPeriod = FALSE,
  minCellCount = 5
)
}
\arguments{
\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not truly support temp tables. To emulate temp 
tables, provide a schema with write privileges where temp tables can be created.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{vocabularyDatabaseSchema}{Schema name where your OMOP vocabulary data resides. This is 
commonly the same as cdmDatabaseSchema. Note that for 
SQL Server, this should include both the database and
schema name, for example 'vocabulary.dbo'.}

\item{cohorts}{A dataframe object with required fields cohortId, sql, json, cohortName}

\item{cohortIds}{Optionally, provide a subset of cohort IDs to restrict the
diagnostics to.}

\item{cohortDatabaseSchema}{Schema name where your cohort table resides. Note that for SQL Server,
this should include both the database and schema name, for example
'scratch.dbo'.}

\item{cohortTable}{Name of the cohort table.}

\item{keep2BillionConceptId}{(Optional) Default FALSE. Do you want to keep concept id above 2 billion.
Per OMOP conventions any conceptId >= 2 billion are considered site specific
custom value that are not shipped as part of default OMOP vocabulary tables.}

\item{runConceptSetOptimization}{Generate and export optimized concept set expression?}

\item{runExcludedConceptSet}{Generate and export excluded concept id in concept set expression?}

\item{runOrphanConcepts}{Generate and export potential orphan concepts?}

\item{runBreakdownIndexEvents}{Generate and export breakdown of index events?}

\item{runBreakdownIndexEventRelativeDays}{(optional) array of days to offset breakdown of index events. 
default value is 0 i.e. no offsets. Options include c(-5:5) to calculate a 
range of days starting -5 of cohort start to +5 of cohort start.}

\item{runIndexDateConceptCoOccurrence}{Generate concept co-occurrence on index date}

\item{runStandardToSourceMappingCount}{Generate and export counts for mapping between standard to non standard as observed in data.}

\item{runConceptCount}{Generate and export count of concept id}

\item{runConceptCountByCalendarPeriod}{Do you want to stratify the counts by calendar period like calendar year, calendar month?}

\item{minCellCount}{The minimum cell count for fields contains person counts or fractions.}
}
\description{
Runs concept set diagnostics on a set of cohorts. For index event breakdown,
the cohorts need to be instantiated.
}
