% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runIncludedSourceConcepts.R
\name{runIncludedSourceConcepts}
\alias{runIncludedSourceConcepts}
\title{Generate and export the source concepts included in the cohorts}
\usage{
runIncludedSourceConcepts(
  connection,
  cohortDefinitionSet,
  tempEmulationSchema,
  cdmDatabaseSchema,
  databaseId,
  exportFolder,
  minCellCount,
  incremental = FALSE,
  incrementalFolder = exportFolder
)
}
\arguments{
\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package.}

\item{cohortDefinitionSet}{A data.frame with cohort definitions created by 
`CohortGenerator::getCohortDefinitionSet` that must include 
the columns cohortId, cohortName, json, sql.}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not truly support temp tables. To emulate temp 
tables, provide a schema with write privileges where temp tables can be created.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{databaseId}{A short string for identifying the database (e.g. 'Synpuf').}

\item{exportFolder}{The folder where the results will be exported to}

\item{minCellCount}{The minimum cell count for fields contains person counts or fractions}

\item{incremental}{`TRUE` or `FALSE` (default). If TRUE diagnostics for cohorts in the 
cohort definition set that have not changed will be skipped and existing results 
csv files will be updated. If FALSE then diagnostics for all cohorts in the cohort 
definition set will be executed and pre-existing results files will be deleted.}

\item{incrementalFolder}{If \code{incremental = TRUE}, specify a folder where records are kept
of which cohort diagnostics has been executed.}
}
\value{
None, it will write csv files to disk.
}
\description{
Provides the amount of distinct subjects and the frequency of the combination
of source concept ids and concept ids listed in the cohorts table,
for which the start and end of the relevant event that occurred to the subjects,
is within their defined observation period. Results are organised per cohort id,
concept set id, concept id and source concept id.
}
