% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTemporalCohortCharacterization.R
\name{runTemporalCohortCharacterization}
\alias{runTemporalCohortCharacterization}
\title{Title}
\usage{
runTemporalCohortCharacterization(
  connection,
  databaseId,
  exportFolder,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTable,
  covariateSettings,
  tempEmulationSchema,
  cdmVersion,
  cohorts,
  cohortCounts,
  minCellCount,
  instantiatedCohorts,
  incremental,
  recordKeepingFile,
  task = "runTemporalCohortCharacterization",
  jobName = "Temporal Cohort characterization",
  covariateValueFileName = file.path(exportFolder, "temporal_covariate_value.csv"),
  covariateValueContFileName = file.path(exportFolder,
    "temporal_covariate_value_dist.csv"),
  covariateRefFileName = file.path(exportFolder, "temporal_covariate_ref.csv"),
  analysisRefFileName = file.path(exportFolder, "temporal_analysis_ref.csv"),
  timeRefFileName = file.path(exportFolder, "temporal_time_ref.csv"),
  minCharacterizationMean = 0.001,
  batchSize = getOption("CohortDiagnostics-FE-batch-size", default = 5)
)
}
\arguments{
\item{connection}{}

\item{databaseId}{}

\item{exportFolder}{}

\item{cdmDatabaseSchema}{}

\item{cohortDatabaseSchema}{}

\item{cohortTable}{}

\item{covariateSettings}{}

\item{tempEmulationSchema}{}

\item{cdmVersion}{}

\item{cohorts}{}

\item{cohortCounts}{}

\item{minCellCount}{}

\item{instantiatedCohorts}{}

\item{incremental}{}

\item{recordKeepingFile}{}

\item{task}{}

\item{jobName}{}

\item{covariateValueFileName}{}

\item{covariateValueContFileName}{}

\item{covariateRefFileName}{}

\item{analysisRefFileName}{}

\item{timeRefFileName}{}

\item{minCharacterizationMean}{}

\item{batchSize}{}
}
\description{
Title
}
