% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTimeSeries.R
\name{runTimeSeries}
\alias{runTimeSeries}
\title{Given a set of instantiated cohorts get time series for the cohorts.}
\usage{
runTimeSeries(
  connection,
  tempEmulationSchema,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTable,
  cohortDefinitionSet,
  runCohortTimeSeries = TRUE,
  runDataSourceTimeSeries = FALSE,
  databaseId,
  exportFolder,
  minCellCount,
  instantiatedCohorts,
  incremental,
  recordKeepingFile,
  observationPeriodDateRange,
  batchSize = getOption("CohortDiagnostics-TimeSeries-batch-size", default = 20)
)
}
\arguments{
\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not truly support temp tables. To emulate temp 
tables, provide a schema with write privileges where temp tables can be created.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{cohortDatabaseSchema}{Schema name where your cohort table resides. Note that for SQL Server,
this should include both the database and schema name, for example
'scratch.dbo'.}

\item{cohortTable}{Name of the cohort table.}

\item{cohortDefinitionSet}{A data.frame with cohort definitions created by 
`CohortGenerator::getCohortDefinitionSet` that must include 
the columns cohortId, cohortName, json, sql.}

\item{runCohortTimeSeries}{Generate and export the cohort level time series?}

\item{runDataSourceTimeSeries}{Generate and export the Data source level time series? i.e.
using all persons found in observation period table.}

\item{databaseId}{The databaseId to retrieve the results for. This is a 
character field values from the `databaseId` field of the `database` table 
of the results data model.}

\item{exportFolder}{The folder where the results will be exported to}

\item{minCellCount}{The minimum cell count for fields contains person counts or fractions}

\item{instantiatedCohorts}{}

\item{incremental}{Create only cohort diagnostics that haven't been created before?}

\item{observationPeriodDateRange}{}

\item{batchSize}{}

\item{databaseIds}{A vector one or more databaseIds to retrieve the results for. This is a 
character field values from the `databaseId` field of the `database` table 
of the results data model.}
}
\value{
None, it will write the results to a csv file
}
\description{
This function first generates a calendar period table, that has
calendar intervals between the \code{timeSeriesMinDate} and \code{timeSeriesMaxDate}.
Calendar Month, Quarter and year are supported.
For each of the calendar interval, time series data are computed. 

Data Source time series: computes time series at the data source level i.e. observation
period table. This output is NOT limited to individuals in the cohort table
but is for ALL people in the datasource (i.e. present in observation period table)
}
