% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APACStudyFudan2025.R
\docType{package}
\name{APACStudyFudan2025-package}
\alias{APACStudyFudan2025}
\alias{APACStudyFudan2025-package}
\title{APACStudyFudan2025: A Study Package for Generating Cohort Diagnostics}
\description{
A package that generates cohort diagnostics for cohort definitions included in the package.
}
\keyword{internal}
