% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Private.R
\name{timeExecution}
\alias{timeExecution}
\title{Internal utility function for logging execution of variables}
\usage{
timeExecution(
  exportFolder,
  taskName,
  cohortIds = NULL,
  parent = NULL,
  start = NA,
  execTime = NA,
  expr = NULL
)
}
\description{
Internal utility function for logging execution of variables
}
